### 2020-08-26 #2603 @Kevin-Mattheus-Moerman
@mtsch can you work on the following:

- [ ] Can you check/fix those missing/invalid DOI's :point_up:? You can call `@whedon check references` to check these here
- [ ] As recommended in [our guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) can you rework/add text to have a `Statement of Need` section in your paper? You can call `@whedon generate pdf` to update the paper pdf here
