### 2020-08-27 f48eadb
Joss paper update (#90)
* add dois to papers, remove unused references

* rework paper to include a statement of need

* move figure to top of page

* capitalize "Need"
