+++ joss/paper.bib
@@ -1,11 +1,13 @@
-@article{edelsbrunner2008persistent,
-  title={Persistent homology-a survey},
-  author={Edelsbrunner, Herbert and Harer, John},
-  journal={Contemporary mathematics},
-  volume={453},
-  pages={257--282},
-  year={2008},
-  publisher={Providence, RI: American Mathematical Society}
+@article{adams2017persistence,
+  title={Persistence images: A stable vector representation of persistent homology},
+  author={Adams, Henry and Emerson, Tegan and Kirby, Michael and Neville, Rachel and Peterson, Chris and Shipman, Patrick and Chepushtanova, Sofya and Hanson, Eric and Motta, Francis and Ziegelmeier, Lori},
+  journal={The Journal of Machine Learning Research},
+  volume={18},
+  number={1},
+  pages={218--252},
+  year={2017},
+  publisher={JMLR. org},
+  doi={10.5555/3122009.3122017}
 }
 @inproceedings{asaad2017topological,
   title={Topological data analysis for image tampering detection},
@@ -13,23 +15,15 @@
   booktitle={International Workshop on Digital Watermarking},
   pages={136--146},
   year={2017},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-319-64185-0_11}
 }
-@article{lee2017quantifying,
-  title={Quantifying similarity of pore-geometry in nanoporous materials},
-  author={Lee, Yongjin and Barthel, Senja D and D{\l}otko, Pawe{\l} and Moosavi, S Mohamad and Hess, Kathryn and Smit, Berend},
-  journal={Nature communications},
-  volume={8},
-  pages={15396},
-  year={2017},
-  publisher={Nature Publishing Group}
-}
-@inproceedings{tralie2016high,
-  title={High-dimensional geometry of sliding window embeddings of periodic videos},
-  author={Tralie, Christopher},
-  booktitle={32nd International Symposium on Computational Geometry (SoCG 2016)},
-  year={2016},
-  organization={Schloss Dagstuhl-Leibniz-Zentrum fuer Informatik}
+@article{bauer2019ripser,
+  title={Ripser: efficient computation of Vietoris-Rips persistence barcodes},
+  author={Bauer, Ulrich},
+  journal={arXiv preprint arXiv:1908.02518},
+  year={2019},
+  url={https://arxiv.org/abs/1908.02518}
 }
 @article{bernoff2016biological,
   title={Biological Aggregation Driven by Social and Environmental Factors: A Nonlocal Model and Its Degenerate Cahn--Hilliard Approximation},
@@ -39,26 +33,57 @@
   number={3},
   pages={1528--1562},
   year={2016},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/15M1031151}
 }
-@article{bauer2019ripser,
-  title={Ripser: efficient computation of Vietoris-Rips persistence barcodes},
-  author={Bauer, Ulrich},
-  journal={arXiv preprint arXiv:1908.02518},
-  year={2019}
+@article{chung2019persistence,
+  title={Persistence curves: A canonical framework for summarizing persistence diagrams},
+  author={Chung, Yu-Min and Lawson, Austin},
+  journal={arXiv preprint arXiv:1904.07768},
+  year={2019},
+  url={https://arxiv.org/abs/1904.07768}
 }
-@article{zhang2020gpu,
-  title={GPU-Accelerated Computation of Vietoris-Rips Persistence Barcodes},
-  author={Zhang, Simon and Xiao, Mengbai and Wang, Hao},
-  journal={arXiv preprint arXiv:2003.07989},
+@article{edelsbrunner2008persistent,
[…]
