+++ joss/paper.md
@@ -10,64 +10,61 @@ authors:
 affiliations:
  - name: Independent Researcher
    index: 1
-date: 24 August 2020
+date: 27 August 2020
 bibliography: paper.bib
 ---
 
-# Summary
+# Introduction
 
 Persistent homology [@edelsbrunner2008persistent] is a relatively recent computational
 technique that extracts topological information from various kinds of datasets. This
 topological information gives us a good overview of the global shape of the data as well as
-giving us a description of its local geometry. Since its introduction it has been used in a
-diverse range of applications, including biology [@bernoff2016biological], material
-science [@lee2017quantifying] to signal processing [@tralie2016high], and computer
-vision [@asaad2017topological]. A problem persistent homology faces is the very large size of
+giving us a description of its local geometry. Since its introduction, it has been used in a
+diverse range of applications, including biology [@bernoff2016biological], material science
+[@lee2017quantifying], signal processing [@tralie2016high], and computer vision
+[@asaad2017topological]. A problem persistent homology faces is the very large size of
 combinatorial structures it has to work with. Recent algorithmic advances employ various
 computational shortcuts to overcome this problem.
 
-Among the most successful implementations of persistent homology is
-Ripser [@bauer2019ripser]. With its speed and low memory usage, it makes persistent homology
+Among the most successful implementations of persistent homology is Ripser
+[@bauer2019ripser]. With its speed and low memory usage, it makes persistent homology
 practical for larger datasets, even in higher dimensions. The introduction of Ripser has
-spawned a whole cottage industry of extensions and wrappers. Some examples include
-Ripser++ [@zhang2020gpu], Lock-free Ripser [@morozov2020towards], Ripser.py [@tralie2018ripser]
-and Cubical Ripser [@kaji2020cubical].
+spawned a whole cottage industry of extensions and wrappers. Some examples include Ripser++
+[@zhang2020gpu], Lock-free Ripser [@morozov2020towards], Ripser.py [@tralie2018ripser] and
+Cubical Ripser [@kaji2020cubical].
 
-In this work, we present Ripserer.jl, a pure Julia implementation of persistent homology
-based on the algorithm that powers Ripser. The main goal of this project is to provide a
-highly generic and extensible persistent homology library that is easy to use for
-non-experts without sacrificing too much performance.
+# Statement of Need
 
-# Package Details
+A significant hurdle in developing new approaches to persistent homology stems from the fact
+that developing an efficient implementation of its matrix reduction algorithm is nontrivial.
 
-Along with its companion package PersistenceDiagrams.jl, Ripserer.jl provides a featureful
-environment for computing persistent homology and integrating it with the rest of Julia's
-data science stack. At the time of writing, it offers the following features.
+To solve this problem, we introduce Ripserer.jl, a pure Julia implementation of persistent
+homology based on the algorithm that powers Ripser. It provides users with an intuitive
+user interface and is readily useful as a topological data analysis framework. The other
+main feature Ripserer.jl provides is the ability to hook into its algorithm through an
+API. This allows researchers to experiment with different approaches to persistent homology
+without having to reimplement the algorithm from scratch or forking an existing repository.
+
+# Summary
+
+Along with its companion package PersistenceDiagrams.jl[^1], Ripserer.jl provides a
+featureful environment for computing persistent homology and integrating it with the rest of
+Julia's data science stack. At the time of writing, it offers the following features.
+
+![Example visualizations. The plot on the left shows the three main representative cocycles
+in the data. The right plot shows the persistence diagram.](figure.png)
 
[…]
