### 2020-08-27 ad42623
remove invalid doi (#91)
