+++ joss/paper.bib
@@ -7,7 +7,7 @@
   pages={218--252},
   year={2017},
   publisher={JMLR. org},
-  doi={10.5555/3122009.3122017}
+  url={https://dl.acm.org/doi/10.5555/3122009.3122017},
 }
 @inproceedings{asaad2017topological,
   title={Topological data analysis for image tampering detection},
