### 2020-08-27 #2603 @VivianePons
@mtsch thank you for fixing the DOI

do you have any recommendations for potential reviewers? You can look at [our list of potential reviewers](https://bit.ly/joss-reviewers) but also suggest names outside the list
