### 2020-08-27 #2603 @mtsch
I looked through the list and the intersection between topological data analysis and Julia appears to be empty.

Maybe @sauln might find this project interesting? Potentially @juliohm as well?
