### 2020-08-27 8447dbb
basic geodesic rips implementation
