+++ test/filtrations/geodesicrips.jl
@@ -1,0 +1,211 @@
+using LightGraphs
+using Ripserer
+using Test
+
+using Ripserer: CircumSimplex, circum, dist
+
+@testset "Circumsimplex" begin
+    @testset "Constructors" begin
+        @test CircumSimplex{1}([2, 1], 1) == CircumSimplex{1}(1, 1, 2)
+        @test CircumSimplex{1}((1, 3), 0f0) == CircumSimplex{1}(2, 0f0, 1f0)
+        @test CircumSimplex{2}([5, 2, 1], 1.0, 2.0) == CircumSimplex{2}(5, 1.0, 2.0)
+    end
+
+    @testset "AbstractSimplex interface" begin
+        for D in (0, 2),
+            T in (Float32, Int),
+            I in (Int32, Int64)
+            @testset "CircumSimplex{$D, $T, $I}" begin
+                b = T(9)
+                c = T(18)
+                i = I(100)
+                sx = CircumSimplex{D}(i, b, c)
+
+                @testset "Type stuff" begin
+                    @test sx ≡ CircumSimplex{D, T, I}(i, b, c)
+                    @test typeof(sx) ≡ CircumSimplex{D, T, I}
+                    D > 0 && @test_throws DomainError CircumSimplex{-D}(i, b, c)
+                    @test eltype(sx) == I
+                end
+
+                @testset "Getters" begin
+                    @test index(sx) == i
+                    @test index(-sx) == i
+                    @test birth(sx) ≡ b
+                    @test circum(sx) ≡ c
+                end
+
+                @testset "Equality, hashing" begin
+                    @test sx == sx
+                    @test sx != CircumSimplex{D + 1}(i, b, c)
+                    @test sx == CircumSimplex{D}(i, b + 1, c)
+                    @test isequal(sx, CircumSimplex{D}(i, b + 1, c))
+                    @test hash(sx) == hash(CircumSimplex{D}(i, b, c + 1))
+                    @test hash(sx) == hash(index(sx))
+                end
+
+                @testset "Signs" begin
+                    @test +sx === sx
+                    @test sx == -sx
+                    @test sx ≡ -(-sx)
+                    @test sign(sx) == 1
+                    @test sign(-sx) == -1
+                    @test abs(sx) ≡ sx
+                    @test abs(-sx) ≡ sx
+                    @test dim(sx) == D
+                end
+
+                @testset "Ordering" begin
+                    @test sx < CircumSimplex{D}(I(i + 1), b + 1, c)
+                    @test sx > CircumSimplex{D}(I(i - 1), b - 1, c)
+
+                    @test sx < CircumSimplex{D}(I(i - 1), b, c)
+                    @test sx > CircumSimplex{D}(I(i + 1), b, c)
+
+                    @test sx < CircumSimplex{D}(i, b, c + 1)
+                    @test sx > CircumSimplex{D}(i, b, c - 1)
+                end
+
+                @testset "Array interface, vertices" begin
+                    verts = vertices(sx)
+
+                    @test eltype(sx) == eltype(verts)
+                    @test length(sx) == length(verts) == D + 1
+                    @test size(sx) == (D + 1,)
+                    @test firstindex(sx) == 1
+                    @test lastindex(sx) == D + 1
+
+                    @test CircumSimplex{D}(verts, birth(sx), circum(sx)) ≡ sx
+
+                    for (i, v) in enumerate(sx)
+                        @test v == verts[i]
+                    end
+
+                    @test begin @inferred vertices(sx); true end
+                end
+
[…]
