### 2020-08-28 a84a84c
add optional apparent pair finding step
