### 2020-08-29 064a9e7
disable for homology, make it work correctly when apparent interval length is nonzero
