### 2020-08-29 #2603 @juliohm
@VivianePons I remember JOSS had a check list on each submission. Is it still available? I remember the bot was responsible for creating this checklist. The review process already started?
