### 2020-08-30 2476b1e
Add information about other Julia packages (#95)
