+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"generated/cocycles/","page":"Representative Cocycles","title":"Representative Cocycles","text":"EditURL = \"https://github.com/mtsch/Ripserer.jl/blob/master/docs/src/examples/cocycles.jl\"","category":"page"},{"location":"generated/cocycles/#Representative-Cocycles","page":"Representative Cocycles","title":"Representative Cocycles","text":"","category":"section"},{"location":"generated/cocycles/","page":"Representative Cocycles","title":"Representative Cocycles","text":"In this example, we will demonstrate how to compute and visualize representative cocycles of persistent homology generators.","category":"page"},{"location":"generated/cocycles/","page":"Representative Cocycles","title":"Representative Cocycles","text":"As always, we start by importing the relevant packages.","category":"page"},{"location":"generated/cocycles/","page":"Representative Cocycles","title":"Representative Cocycles","text":"using Ripserer\nusing Plots\ngr(); nothing # hide","category":"page"},{"location":"generated/cocycles/","page":"Representative Cocycles","title":"Representative Cocycles","text":"We define a function that generates some data sampled from a curve.","category":"page"},{"location":"generated/cocycles/","page":"Representative Cocycles","title":"Representative Cocycles","text":"function curve(n)\n    [(sin(t)+t/10, cos(t)+t/10) for t in range(0, 2π, length=n)]\nend\n\nplot(curve(100), legend=false, title=\"Curve\", aspect_ratio=1, xlab=\"x\", ylab=\"y\")","category":"page"},{"location":"generated/cocycles/","page":"Representative Cocycles","title":"Representative Cocycles","text":"Then, we compute the persistent homology of that data with various numbers of points. We plot the representative cocycles of each result and turn them into an animation. This data set will always have at most one one-dimensional class. We invoke plot(interval, data) to plot the representative cocycle of the class. The same plot could be created by invoking representative(interval). We use the threshold_strict argument to only plot the simplices with diameter strictly lower than the death time of the interval.","category":"page"},{"location":"generated/cocycles/","page":"Representative Cocycles","title":"Representative Cocycles","text":"anim = @animate for i in vcat(3:100, 100:-1:3)\n    points = curve(i)\n    res = ripserer(points, reps=true)\n\n    plt1 = plot(title=\"1-dimensional Representative Cocycle\",\n                xlab=\"x\",\n                ylab=\"y\",\n                legend=false,\n                ylim=(-0.8, 1.8),\n                aspect_ratio=1)\n    if length(res[2]) > 0\n        interval = res[2][1]\n        plot!(plt1, interval, points,\n              alpha=0.2,\n              linewidth=2,\n              threshold_strict=death(interval),\n              color=3)\n    end\n    scatter!(plt1, points, color=2)\n    plot(plt1, barcode(res), plot(res),\n         layout=@layout([a [b; c]]),\n         size=(800, 600))\nend\nmp4(anim, \"cocycles_anim.mp4\") # hide","category":"page"},{"location":"generated/cocycles/","page":"Representative Cocycles","title":"Representative Cocycles","text":"We can save the animation by running the following.","category":"page"},{"location":"generated/cocycles/","page":"Representative Cocycles","title":"Representative Cocycles","text":"mp4(anim, \"cocycles_anim.mp4\")","category":"page"},{"location":"generated/cocycles/","page":"Representative Cocycles","title":"Representative Cocycles","text":"","category":"page"},{"location":"generated
[…]
