### 2020-08-30 5105d61
fix buggy test, make sure it doesn't happen again
