+++ test/filtrations/new-filtrations.jl
@@ -7,6 +7,7 @@ using StaticArrays
 using Test
 
 include(joinpath(@__DIR__, "test-datasets.jl"))
+using Ripserer: nv, edges
 
 # Julia 1.0 does not allow these to be defined inside @testset.
 struct NewFiltration <: Ripserer.AbstractFiltration{Int, Int} end
@@ -53,19 +54,23 @@ end
 # This test is mostly supposed to test that pre-finding apparent pairs with
 # `find_apparent_pairs` first works correctly. On the side, it also tests that rips
 # filtrations that only overload `dist` and `threshold` work fine.
-struct ApparentPairsRips{I, T} <: Ripserer.AbstractRipsFiltration{I, T}
-    rips::Rips{I, T}
+struct ApparentPairsRips{I, T, R<:Rips{I, T}} <: Ripserer.AbstractRipsFiltration{I, T}
+    rips::R
+    apparent::Ref{Int}
+    normal::Ref{Int}
 end
 
-ApparentPairsRips(data; kwargs...) = ApparentPairsRips(Rips(data; kwargs...))
+function ApparentPairsRips(data; kwargs...)
+    return ApparentPairsRips(Rips(data; kwargs...), Ref(0), Ref(0))
+end
 
 Ripserer.dist(rw::ApparentPairsRips) = Ripserer.dist(rw.rips)
 Ripserer.threshold(rw::ApparentPairsRips) = Ripserer.threshold(rw.rips)
 
-# This works fine but is slower than the original algorithm as apparent pairs are found in
-# `initialize_coboundary!` anyway. Doing this in parallel or on the GPU is an option,
-# however.  See https://arxiv.org/abs/2003.07989
-function Ripserer.find_apparent_pairs(rw::ApparentPairsRips, columns, is_cohomology, _)
+# This works fine but is slower than the original algorithm (even if the pair finding code
+# was optimized) as apparent pairs are found in `initialize_coboundary!` anyway. Doing this
+# in parallel or on the GPU is an option, however.  See https://arxiv.org/abs/2003.07989
+function Ripserer.find_apparent_pairs(rw::ApparentPairsRips, columns, _)
     S = eltype(columns)
     C = Ripserer.simplex_type(rw, dim(S)+1)
     cols_left = S[]
@@ -74,8 +79,10 @@ function Ripserer.find_apparent_pairs(rw::ApparentPairsRips, columns, is_cohomol
         τ = minimum(Ripserer.coboundary(rw, σ)) # This is broken if coboundary is empty.
         σ′ = maximum(Ripserer.boundary(rw, τ))
         if σ′ == σ
+            rw.apparent[] += 1
             push!(apparent, (σ, τ))
         else
+            rw.normal[] += 1
             push!(cols_left, σ)
         end
     end
@@ -85,16 +92,21 @@ end
 @testset "Rips with `find_apparent_pairs` overloaded." begin
     for m in (2, 3, 5), t in (2, 1)
         @testset "projective plane with modulus=$m, threshold=$t" begin
+            appa = ApparentPairsRips(projective_plane; threshold=t)
             res_rips = ripserer(Rips(projective_plane; threshold=t); modulus=m)
-            res_appa = ripserer(ApparentPairsRips(projective_plane; threshold=t); modulus=m)
+            res_appa = ripserer(appa; modulus=m)
             @test res_rips == res_appa
+            # The nv - 1 edges were cleared by the twist algorithm.
+            @test appa.apparent[] + appa.normal[] + nv(appa) - 1 == length(edges(appa))
         end
     end
     for m in (2, 3, 5), t in (8, 4)
[…]
