### 2020-08-30 12ff269
Add optional apparent pair finding step (#93)
* add optional apparent pair finding step

* disable for homology, make it work correctly when apparent interval length is nonzero

* typo in comment

* bump version

* fix buggy test, make sure it doesn't happen again
