+++ test/filtrations/new-filtrations.jl
@@ -2,20 +2,14 @@ using Compat
 using Distances
 using PersistenceDiagrams
 using Ripserer
+using SparseArrays
 using StaticArrays
 using Test
 
+include(joinpath(@__DIR__, "test-datasets.jl"))
+using Ripserer: nv, edges
+
 # Julia 1.0 does not allow these to be defined inside @testset.
-struct NewRips <: Ripserer.AbstractRipsFiltration{Int, Float64} end
-Ripserer.dist(::NewRips) = Float64[i ≠ j for i in 1:10, j in 1:10]
-Ripserer.dist(::NewRips, i, j) = 1.0
-
-@testset "New rips filtration" begin
-    d0, d1 = ripserer(NewRips())
-    @test d0 == [fill((0.0, 1.0), 9); (0.0, Inf)]
-    @test d1 == []
-end
-
 struct NewFiltration <: Ripserer.AbstractFiltration{Int, Int} end
 
 function Ripserer.unsafe_simplex(
@@ -55,4 +49,110 @@ end
     d0, d1 = ripserer(NewFiltration(), reps=true)
     @test d0 == []
     @test d1 == []
+end
+
+# This test is mostly supposed to test that pre-finding apparent pairs with
+# `find_apparent_pairs` first works correctly. On the side, it also tests that rips
+# filtrations that only overload `dist` and `threshold` work fine.
+struct ApparentPairsRips{I, T, R<:Rips{I, T}} <: Ripserer.AbstractRipsFiltration{I, T}
+    rips::R
+    apparent::Ref{Int}
+    normal::Ref{Int}
+end
+
+function ApparentPairsRips(data; kwargs...)
+    return ApparentPairsRips(Rips(data; kwargs...), Ref(0), Ref(0))
+end
+
+Ripserer.dist(rw::ApparentPairsRips) = Ripserer.dist(rw.rips)
+Ripserer.threshold(rw::ApparentPairsRips) = Ripserer.threshold(rw.rips)
+
+# This works fine but is slower than the original algorithm (even if the pair finding code
+# was optimized) as apparent pairs are found in `initialize_coboundary!` anyway. Doing this
+# in parallel or on the GPU is an option, however.  See https://arxiv.org/abs/2003.07989
+function Ripserer.find_apparent_pairs(rw::ApparentPairsRips, columns, _)
+    S = eltype(columns)
+    C = Ripserer.simplex_type(rw, dim(S)+1)
+    cols_left = S[]
+    apparent = Tuple{S, C}[]
+    for σ in columns
+        τ = minimum(Ripserer.coboundary(rw, σ)) # This is broken if coboundary is empty.
+        σ′ = maximum(Ripserer.boundary(rw, τ))
+        if σ′ == σ
+            rw.apparent[] += 1
+            push!(apparent, (σ, τ))
+        else
+            rw.normal[] += 1
+            push!(cols_left, σ)
+        end
+    end
+    return cols_left, apparent
+end
+
+@testset "Rips with `find_apparent_pairs` overloaded." begin
+    for m in (2, 3, 5), t in (2, 1)
+        @testset "projective plane with modulus=$m, threshold=$t" begin
+            appa = ApparentPairsRips(projective_plane; threshold=t)
+            res_rips = ripserer(Rips(projective_plane; threshold=t); modulus=m)
+            res_appa = ripserer(appa; modulus=m)
+            @test res_rips == res_appa
+            # The nv - 1 edges were cleared by the twist algorithm.
+            @test appa.apparent[] + appa.normal[] + nv(appa) - 1 == length(edges(appa))
+        end
+    end
[…]
