### 2020-08-31 #2603 @VivianePons
The checklist appear on the review issue, we are still on the "pre review". I have just confirmed a second reviewer and am about to launch the review :)
