### 2020-08-31 #2614 @whedon
[REVIEW]: Ripserer.jl: flexible and efficient persistent homology computation in Julia
**Submitting author:** @mtsch (<a href="http://orcid.org/0000-0003-0734-2719">Matija Čufar</a>)
**Repository:** <a href="https://github.com/mtsch/Ripserer.jl" target ="_blank">https://github.com/mtsch/Ripserer.jl</a>
**Version:** v0.15.2
