+++ joss/paper.bib
@@ -50,7 +50,8 @@
   volume={453},
   pages={257--282},
   year={2008},
-  publisher={Providence, RI: American Mathematical Society}
+  publisher={Providence, RI: American Mathematical Society},
+  doi={10.1090/conm/453/08802}
 }
 @article{kaji2020cubical,
   title={Cubical Ripser: Software for computing persistent homology of image and volume data},
