### 2020-08-31 #2614 @juliohm
I love to review submissions such as this, really well-written, to the point, and a great piece of software! Thanks @mtsch for this amazing work. I only have one minor comment that I already shared as a GitHub issue yesterday about the related Julia packages you recently added to the docs: https://mtsch.github.io/Ripserer.jl/dev/related-work/

Do you think that these should also be listed in the JOSS paper or at least cited in the list of references? I noticed that you cited the major alternatives from other languages, but perhaps it would be a good idea to list related work in Julia.
