### 2020-09-02 fb76800
JOSS: Add new references, minor grammar fixes
