+++ joss/Makefile
@@ -1,2 +1,4 @@
-paper:
+pdf:
 	pandoc --filter pandoc-citeproc paper.md -s -o paper.pdf
+html:
+	pandoc --filter pandoc-citeproc paper.md -s -o paper.html
