### 2020-09-02 594a76a
JOSS: Add more references (#97)
* JOSS: Add new references, minor grammar fixes

* move image back to top of page
