+++ joss/paper.bib
@@ -113,3 +113,28 @@
   publisher={SIAM},
   doi={10.1137/141000671}
 }
+
+@article{henselman2016matroid,
+  title={Matroid filtrations and computational persistent homology},
+  author={Henselman, Gregory and Ghrist, Robert},
+  journal={arXiv preprint arXiv:1606.00199},
+  year={2016}
+}
+
+@article{brehm2018sparips,
+  title={Sparips},
+  author={Brehm, Bernhard and Hardering, Hanne},
+  journal={arXiv preprint arXiv:1807.09982},
+  year={2018}
+}
+@article{lutgehetmann2020computing,
+  title={Computing persistent homology of directed flag complexes},
+  author={L{\"u}tgehetmann, Daniel and Govc, Dejan and Smith, Jason P and Levi, Ran},
+  journal={Algorithms},
+  volume={13},
+  number={1},
+  pages={19},
+  year={2020},
+  publisher={Multidisciplinary Digital Publishing Institute},
+  doi={10.3390/a13010019}
+}
