### 2020-09-02 #2614 @mtsch
Thanks for the kind words!

I added a short paragraph listing Julia packages. I also added citation for Julia and a reference to a Ripser fork I forgot to include.
