### 2020-09-08 5d55035
improved implementation
* use A* instead of Floyd Warshall
* add `dist` for all filtrations
* tests
