+++ src/cycles.jl
@@ -1,40 +1,48 @@
-struct OneSkeleton{T, F<:AbstractFiltration, S<:AbstractSimplex{1, T}} <: AbstractGraph{Int}
+struct OneSkeleton{T, F<:AbstractFiltration, S<:AbstractSimplex{1}} <: AbstractGraph{Int}
     filtration::F
     threshold::T
     removed::Set{S}
 end
-function OneSkeleton(filtration, threshold, removed=())
+function OneSkeleton(
+    filtration::F, thresh::T=threshold(filtration), removed=()
+) where {F, T}
     S = simplex_type(filtration, 1)
     removed = Set{S}(removed)
-    return OneSkeleton(filtration, threshold, removed)
+    return OneSkeleton{T, F, S}(filtration, thresh, removed)
 end
 
+_birth_or_value(σ::AbstractSimplex) = birth(σ)
+_birth_or_value(σ) = σ
 _in(σ::S, g::OneSkeleton{S}) where S = !isnothing(σ) && σ ≤ g.threshold && σ ∉ g.removed
 _in(σ, g::OneSkeleton) = !isnothing(σ) && birth(σ) ≤ g.threshold && σ ∉ g.removed
 
 LightGraphs.edgetype(::OneSkeleton) = Edge{Int}
 
 function LightGraphs.has_edge(g::OneSkeleton, u::Integer, v::Integer)
-    σ = unsafe_simplex(g.filtration, Val(1), TupleTools.sort((v, u), rev=true))
-    return _in(σ, g)
+    if u ∉ vertices(g) || v ∉ vertices(g)
+        return false
+    else
+        σ = simplex(g.filtration, Val(1), _inv_linear.(Ref(g), (v, u)))
+        return _in(σ, g)
+    end
 end
 
 function LightGraphs.edges(g::OneSkeleton)
     result = edgetype(g)[]
     for sx in Ripserer.edges(g.filtration)
-        if birth(sx) ≤ g.threshold && sx ∉ g.repset
-            u, v = sx
+        if _in(sx, g)
+            u, v = _linear.(Ref(g), sx)
             push!(result, Edge(u, v))
         end
     end
     return result
 end
 function LightGraphs.outneighbors(g::OneSkeleton, u::Integer)
-    root = simplex(g.filtration, Val(0), (u,))
+    root = simplex(g.filtration, Val(0), (_inv_linear(g, u),))
     neighbors = Int[]
     for sx in Ripserer.coboundary(g.filtration, root)
         if _in(sx, g)
-            v, w = sx
+            v, w = _linear.(Ref(g), sx)
             if v == u
                 push!(neighbors, w)
             else
@@ -44,50 +52,86 @@ function LightGraphs.outneighbors(g::OneSkeleton, u::Integer)
     end
     return neighbors
 end
+LightGraphs.inneighbors(g::OneSkeleton, u::Integer) = outneighbors(g, u)
 
 Base.eltype(::OneSkeleton) = Int
 LightGraphs.has_vertex(g::OneSkeleton, u) = 1 ≤ u ≤ nv(g)
 LightGraphs.ne(g::OneSkeleton) = length(edges(g))
-LightGraphs.nv(g::OneSkeleton) = length(nv(g.filtration))
-LightGraphs.vertices(g::OneSkeleton) = vertices(g)
+LightGraphs.nv(g::OneSkeleton) = nv(g.filtration)
+LightGraphs.vertices(g::OneSkeleton) = Base.OneTo(nv(g))
 LightGraphs.is_directed(::OneSkeleton) = false
 LightGraphs.is_directed(::Type{<:OneSkeleton}) = false
 
 # need default dist for filtrations...
[…]
