### 2020-09-12 1af5fa0
Improve progress printing
