### 2020-09-13 0f4b3ce
refactor filtrations
* drop `birth` in favor of `births`
* drop `dist` in favor of `adjacency_matrix`
* drop `SparseRips`
