+++ src/filtrations/rips.jl
@@ -1,15 +1,16 @@
 """
     AbstractRipsFiltration{I<:Signed, T} <: AbstractFiltration{I, T}
 
-An abstract Vietoris-Rips filtration. Its subtypes can only overload [`dist`](@ref) and get
-default implementations for the rest of the filtration interface.
+An abstract Vietoris-Rips filtration. Its subtypes can only overload
+[`adjacency_matrix`](@ref) and get default implementations for the rest of the filtration
+interface.
 
 # Example
 
 ```jldoctest
 julia> struct MyRips <: Ripserer.AbstractRipsFiltration{Int, Float16} end
 
-julia> Ripserer.dist(::MyRips) = [0 1 1; 1 0 1; 1 1 0]
+julia> Ripserer.adjacency_matrix(::MyRips) = [0 1 1; 1 0 1; 1 1 0]
 
 julia> ripserer(MyRips())
 2-element Array{PersistenceDiagrams.PersistenceDiagram,1}:
@@ -20,38 +21,12 @@ julia> ripserer(MyRips())
 """
 abstract type AbstractRipsFiltration{I<:Signed, T} <: AbstractFiltration{I, T} end
 
-"""
-    dist(::AbstractRipsFiltration, u, v)
-    dist(::AbstractRipsFiltration)
-
-Return the distance between vertices `u` and `v`. If the distance is somehow invalid, it may
-return `missing` instead. If `u` and `v` are not given, return the distance matrix.
-
-# Examples
-
-```jldoctest
-julia> flt = Rips([1 1 2; 1 0 1; 2 1 0]);
-
-julia> Ripserer.dist(flt, 1, 3)
-2
-
-julia> Ripserer.dist(flt)
-3×3 Array{Int64,2}:
- 1  1  2
- 1  0  1
- 2  1  0
-
-```
-"""
-dist(rips::AbstractRipsFiltration, i, j) = dist(rips)[i, j]
-
-nv(rips::AbstractRipsFiltration) = size(dist(rips), 1)
-birth(rips::AbstractRipsFiltration) = diag(dist(rips))
+nv(rips::AbstractRipsFiltration) = size(adjacency_matrix(rips), 1)
+births(rips::AbstractRipsFiltration) = diag(adjacency_matrix(rips))
 simplex_type(::Type{<:AbstractRipsFiltration{I, T}}, D) where {I, T} = Simplex{D, T, I}
-adjacency_matrix(rips::AbstractRipsFiltration) = dist(rips)
 
 function edges(rips::AbstractRipsFiltration)
-    if issparse(dist(rips))
+    if issparse(adjacency_matrix(rips))
         _sparse_edges(rips)
     else
         _full_edges(rips)
@@ -60,7 +35,7 @@ end
 
 function _full_edges(rips::AbstractRipsFiltration)
     result = edge_type(rips)[]
-    @inbounds for u in 1:size(dist(rips), 1), v in u+1:size(dist(rips), 1)
+    @inbounds for u in 1:nv(rips), v in u+1:nv(rips)
         sx = unsafe_simplex(rips, Val(1), (v, u), 1)
         !isnothing(sx) && push!(result, sx)
     end
@@ -68,11 +43,11 @@ function _full_edges(rips::AbstractRipsFiltration)
 end
 
 function _sparse_edges(rips::AbstractRipsFiltration)
+    adj = adjacency_matrix(rips)
     result = edge_type(rips)[]
-    rows = rowvals(rips.dist)
-    vals = nonzeros(rips.dist)
-    for u in 1:size(rips.dist, 1)
-        for i in nzrange(dist(rips), u)
+    rows = rowvals(adj)
+    for u in 1:nv(rips)
+        for i in nzrange(adj, u)
[…]
