+++ test/cycles.jl
@@ -1,0 +1,249 @@
+using LightGraphs
+using Ripserer
+using Test
+
+using Ripserer: dist, reconstruct_cycle, OneSkeleton
+
+@testset "OneSkeleton respects the LightGraphs interface" begin
+    @testset "no threshold or removed simplices" begin
+        flt = Rips([
+            0 1 2 3 2 1
+            1 0 1 2 3 2
+            2 1 0 1 2 3
+            3 2 1 0 1 2
+            2 3 2 1 0 1
+            1 2 3 2 1 0
+        ])
+
+        g = OneSkeleton(flt)
+        @test eltype(g) ≡ Int
+        @test nv(g) == 6
+        @test ne(g) == binomial(6, 2)
+        @test has_vertex(g, 1)
+        @test has_vertex(g, 6)
+        @test !has_vertex(g, 7)
+        @test !has_vertex(g, 0)
+        @test edges(g) == [Edge(j, i) for i in 1:5 for j in i+1:6]
+        @test vertices(g) == 1:6
+        @test is_directed(g) == is_directed(typeof(g)) == false
+        @test weights(g) == dist(flt)
+        @test has_edge(g, 1, 2)
+        @test has_edge(g, 3, 6)
+        @test !has_edge(g, 3, 7)
+        @test neighbors(g, 1) == [6, 5, 4, 3, 2]
+    end
+
+    @testset "threshold and removed vertices" begin
+        dists = [
+            0 1 2 3 2 1
+            1 0 1 2 3 2
+            2 1 0 1 2 3
+            3 2 1 0 1 2
+            2 3 2 1 0 1
+            1 2 3 2 1 0
+        ]
+        flt = Rips(dists)
+
+        g = OneSkeleton(flt, 2, [Simplex{1}((2, 1), 1), Simplex{1}((3, 1), 2)])
+        @test eltype(g) ≡ Int
+        @test nv(g) == 6
+        @test ne(g) == binomial(6, 2) - 2 - 3
+        @test has_vertex(g, 1)
+        @test has_vertex(g, 6)
+        @test !has_vertex(g, 7)
+        @test !has_vertex(g, 0)
+        @test edges(g) == [Edge(j, i) for i in 1:5 for j in i+1:6
+                           if dists[i, j] ≠ 3 && (j, i) ∉ [(2, 1), (3, 1)]]
+        @test vertices(g) == 1:6
+        @test is_directed(g) == is_directed(typeof(g)) == false
+        @test weights(g) == dist(flt)
+        @test !has_edge(g, 1, 2)
+        @test !has_edge(g, 3, 6)
+        @test has_edge(g, 3, 5)
+        @test !has_edge(g, 3, 7)
+        @test neighbors(g, 1) == [6, 5]
+    end
+
+    @testset "simplex as threshold" begin
+        dists = [0 1 1 1
+                 1 0 1 1
+                 1 1 0 1
+                 1 1 1 0]
+        flt = Rips(dists)
+        g = OneSkeleton(flt, Simplex{1}((3, 2), 1))
+
+        @test eltype(g) ≡ Int
+        @test nv(g) == 4
+        @test ne(g) == 4
+        @test has_vertex(g, 1)
+        @test has_vertex(g, 4)
+        @test !has_vertex(g, 7)
+        @test !has_vertex(g, 0)
+        @test edges(g) == [Edge(j, i) for i in 1:3 for j in i+1:4 if (j, i) ≥ (3, 2)]
+        @test vertices(g) == 1:4
+        @test is_directed(g) == is_directed(typeof(g)) == false
+        @test weights(g) == dist(flt)
+        @test !has_edge(g, 1, 2)
[…]
