### 2020-09-13 f3559c7
performance fix: use views for births
