### 2020-09-13 4d4baf7
bugfix: copy adjacency matrix before `fkeep!`
