### 2020-09-14 be28584
Filtration refactor (#98)
* Improve progress printing

* refactor filtrations

* drop `birth` in favor of `births`
* drop `dist` in favor of `adjacency_matrix`
* drop `SparseRips`

* bump version

* performance fix: use views for births

* bugfix: copy adjacency matrix before `fkeep!`

* benchmark on Julia v1.5

* fix link in benchmark docs
