+++ test/filtrations/rips.jl
@@ -7,7 +7,7 @@ using StaticArrays
 using Suppressor
 using Test
 
-using Ripserer: distances, vertex_type, edge_type, dist, edges, nv, unsafe_simplex,
+using Ripserer: distances, births, adjacency_matrix, edges, nv, unsafe_simplex,
     ChainElement, PackedElement
 
 include("test-datasets.jl")
@@ -22,93 +22,113 @@ include("test-datasets.jl")
     end
 end
 
-for Filtration in (Rips, SparseRips)
-    @testset "$(string(Filtration))" begin
-        @testset "With no threshold" begin
-            filtration = Filtration(Float64[0 1 2 9; 1 0 3 9; 2 3 0 4; 9 9 4 0])
+for sparse in (true, false)
+    @testset "With no threshold, sparse=$sparse" begin
+        filtration = Rips(Float64[0 1 2 9; 1 0 3 9; 2 3 0 4; 9 9 4 0], sparse=sparse)
 
-            @test sprint(show, filtration) ==
-                "$(nameof(Filtration)){Int64, Float64}(nv=4)"
+        @test sprint(show, filtration) ==
+            "Rips{Int64, Float64}(nv=4, sparse=$sparse)"
 
-            @test nv(filtration) == 4
-            @test dist(filtration, 3, 3) == 0.0
-            @test dist(filtration, 1, 2) == 1.0
-            @test dist(filtration, 1, 3) == 2.0
-            @test dist(filtration, 3, 2) == 3.0
-
-            @test threshold(filtration) == 4.0
-            @test vertex_type(filtration) === Simplex{0, Float64, Int}
-            @test edge_type(filtration) === Simplex{1, Float64, Int}
-            @test birth(filtration, 1) == 0
-            @test birth(filtration, 4) == 0
-
-            @test dist(filtration) == filtration.dist
-
-            @test unsafe_simplex(filtration, Val(0), (1,)) === Simplex{0}(1, 0.0)
-            @test unsafe_simplex(filtration, Val(1), (2, 1), -1) === -Simplex{1}(1, 1.0)
-            @test simplex(filtration, Val(2), (1, 3, 2)) === Simplex{2}(1, 3.0)
-            @test_throws ArgumentError simplex(filtration, Val(2), (1, 1, 2))
+        @test nv(filtration) == 4
+        if issparse(adjacency_matrix(filtration))
+            @test adjacency_matrix(filtration) == [0 1 2 0; 1 0 3 0; 2 3 0 4; 0 0 4 0]
+        else
+            @test adjacency_matrix(filtration) == [0 1 2 9; 1 0 3 9; 2 3 0 4; 9 9 4 0]
         end
-        @testset "With threshold and index type" begin
-            filtration = Filtration{Int128}([1 1 2; 1 2 3; 2 3 2]; threshold=2)
 
-            @test sprint(show, filtration) ==
-                "$(nameof(Filtration)){Int128, Int64}(nv=3)"
+        @test threshold(filtration) == 4.0
+        @test births(filtration) == zeros(4)
 
-            @test nv(filtration) == 3
-            @test threshold(filtration) == 2
-            @test dist(filtration, 3, 3) == 2
-            @test dist(filtration, 1, 2) == 1
-            @test dist(filtration, 1, 3) == 2
-            @test dist(filtration, 3, 2) === (issparse(filtration.dist) ? missing : 3)
-            @test threshold(filtration) == 2
-            @test vertex_type(filtration) === Simplex{0, Int, Int128}
-            @test edge_type(filtration) === Simplex{1, Int, Int128}
+        @test adjacency_matrix(filtration) == filtration.adj
 
[…]
