### 2020-09-14 c5c744e
explain how benchmarks were measured and fix typos
