+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"references/#Acknowledgments","page":"Acknowledgements and References","title":"Acknowledgments","text":"","category":"section"},{"location":"references/","page":"Acknowledgements and References","title":"Acknowledgements and References","text":"I would like to thank:","category":"page"},{"location":"references/","page":"Acknowledgements and References","title":"Acknowledgements and References","text":"@ubauer for creating the original Ripser on which this project is based.\n@ctralie and @sauln for creating ripser.py which has been a source of inspiration.\nŽiga Virk, for giving ideas and helping with the theoretical side of things.","category":"page"},{"location":"references/#References","page":"Acknowledgements and References","title":"References","text":"","category":"section"},{"location":"references/","page":"Acknowledgements and References","title":"Acknowledgements and References","text":"Bauer, U. (2019). Ripser: efficient computation of Vietoris-Rips persistence barcodes. arXiv preprint arXiv:1908.02518.","category":"page"},{"location":"references/","page":"Acknowledgements and References","title":"Acknowledgements and References","text":"Kaji, S., Sudo, T., & Ahara, K. (2020). Cubical Ripser: Software for computing persistent homology of image and volume data. arXiv preprint arXiv:2005.12692.","category":"page"},{"location":"references/","page":"Acknowledgements and References","title":"Acknowledgements and References","text":"Wagner, H., Chen, C., & Vuçini, E. (2012). Efficient computation of persistent homology for cubical data. In Topological methods in data analysis and visualization II (pp. 91-106). Springer, Berlin, Heidelberg.","category":"page"},{"location":"references/","page":"Acknowledgements and References","title":"Acknowledgements and References","text":"Chen, C., & Kerber, M. (2011, March). Persistent homology computation with a twist. In Proceedings 27th European Workshop on Computational Geometry (Vol. 11, pp. 197-200).","category":"page"},{"location":"references/","page":"Acknowledgements and References","title":"Acknowledgements and References","text":"De Silva, V., Morozov, D., & Vejdemo-Johansson, M. (2011). Persistent cohomology and circular coordinates. Discrete & Computational Geometry, 45(4), 737-759.","category":"page"},{"location":"references/","page":"Acknowledgements and References","title":"Acknowledgements and References","text":"Zomorodian, A., & Carlsson, G. (2005). Computing persistent homology. Discrete & Computational Geometry, 33(2), 249-274.","category":"page"},{"location":"references/","page":"Acknowledgements and References","title":"Acknowledgements and References","text":"Edelsbrunner, H. (1993, July). The union of balls and its dual shape. In Proceedings of the ninth annual symposium on Computational geometry (pp. 218-231).","category":"page"},{"location":"generated/cubical/","page":"Cubical Persistent Homology","title":"Cubical Persistent Homology","text":"EditURL = \"https://github.com/mtsch/Ripserer.jl/blob/master/docs/src/examples/cubical.jl\"","category":"page"},{"location":"generated/cubical/#Cubical-Persistent-Homology","page":"Cubical Persistent Homology","title":"Cubical Persistent Homology","text":"","category":"section"},{"location":"generated/cubical/","page":"Cubical Persistent Homology","title":"Cubical Persistent Homology","text":"In this example, we will demonstrate computing sublevel set persistent homology of time series and image data. We will need the following packages.","category":"page"},{"location":"generated/cubical/","page":"Cubical Persistent Homology","title":"Cubical Persistent Homology","text":"using Ripserer\nusing Images\nusing Plots\ngr() # hide\nusing Random\nRandom.seed!(1337)\nnothing # hide","category":"page"},{"location":"generated/cubical/","page":"Cubical Persi
[…]
