### 2020-09-14 9cf58db
fix bug with sparse points and inaccurate euclidean distance
