### 2020-09-14 87ff2c3
only run each benchmark once
