+++ benchmark/bench_rips.jl
@@ -26,7 +26,7 @@ for b in full_benchmarks
         data = load_data(joinpath(@__DIR__, "data", b.file))
         suite[sparse ? "sparse" : "dense"][name] = @benchmarkable ripserer(
             $data; dim_max=$(b.dim_max), threshold=$(b.threshold), sparse=$sparse
-        ) gcsample=true seconds=60
+        ) samples=1
     end
 end
 
@@ -35,7 +35,7 @@ for b in sparse_benchmarks
     data = load_data(joinpath(@__DIR__, "data", b.file))
     suite["sparse"][name] = @benchmarkable ripserer(
         $data; dim_max=$(b.dim_max)
-    ) gcsample=true seconds=60
+    ) samples=1
 end
 
 end
