### 2020-09-14 3e5c61e
Minor fixes (#100)
* fix bug with sparse points and inaccurate euclidean distance

* fix broken test, change `distances`

* test new warning
