### 2020-09-14 5988d88
only build docs on master, only report coverage on current
