### 2020-09-15 e7622c0
Benchmark CI update (#99)
* in docs, explain how benchmarks were measured and fix typos.

* update benchmark ci code.

* only build docs on master, only report coverage on current.

* skip alpha benchmark.
