+++ benchmark/benchmarks.jl
@@ -2,9 +2,12 @@ using Ripserer
 using BenchmarkTools
 
 const SUITE = BenchmarkGroup()
+const SKIP = ["bench_alpha.jl"]
 
 for file in readdir(@__DIR__)
-    if startswith(file, "bench_") && endswith(file, ".jl")
+    if file in SKIP
+        @warn "skipping $file"
+    elseif startswith(file, "bench_") && endswith(file, ".jl")
         SUITE[file[length("bench_") + 1:end - length(".jl")]] = include(file)
     end
 end
