### 2020-09-16 aeff979
fix a_star heuristic
