### 2020-09-17 b2426d0
Add homology benchmarks
