### 2020-09-17 2f6e46d
move computation code to directory, split out cohomology computation
