+++ src/computation/cohomology.jl
@@ -1,0 +1,239 @@
+struct CoboundaryMatrix{T, F, S, R, C}
+    filtration::F
+    #reduced::ReducedMatrix{Cofacet, SimplexElem, O}
+    reduced::R
+    #chain::WorkingChain{CofacetElem, O}
+    chain::C
+    columns_to_reduce::Vector{S}
+    columns_to_skip::Vector{S}
+end
+
+function CoboundaryMatrix(::Type{T}, filtration, columns_to_reduce, columns_to_skip) where T
+    Simplex = eltype(columns_to_reduce)
+    Cofacet = simplex_type(filtration, dim(Simplex) + 1)
+    ordering = Base.Order.Forward
+    SimplexElem = chain_element_type(Simplex, T)
+    CofacetElem = chain_element_type(Cofacet, T)
+
+    reduced = ReducedMatrix{Cofacet, SimplexElem}(ordering)
+    sizehint!(reduced, length(columns_to_reduce))
+    chain = WorkingChain{CofacetElem}(ordering)
+
+    return CoboundaryMatrix{T, typeof(filtration), Simplex, typeof(reduced), typeof(chain)}(
+        filtration, reduced, chain, columns_to_reduce, columns_to_skip
+    )
+end
+
+function coboundary(matrix::CoboundaryMatrix, simplex::AbstractSimplex)
+    return coboundary(matrix.filtration, simplex)
+end
+
+field_type(::CoboundaryMatrix{T}) where T = T
+simplex_type(::CoboundaryMatrix{<:Any, <:Any, S}) where S = S
+simplex_element(::CoboundaryMatrix{T, <:Any, S}) where {T, S} = chain_element_type(S, T)
+dim(cm::CoboundaryMatrix) = dim(simplex_type(cm))
+cofacet_type(cm::CoboundaryMatrix{<:Any, F}) where F = simplex_type(F, dim(cm) + 1)
+cofacet_element(cm::CoboundaryMatrix{T}) where {T} = chain_element_type(cofacet_type(cm), T)
+
+function initialize_coboundary!(matrix::CoboundaryMatrix, column_simplex)
+    empty!(matrix.chain)
+    # Emergent pairs: we are looking for pairs of simplices (σ, τ) where σ is the youngest
+    # facet of τ and τ is the oldest cofacet of σ. These pairs give birth to persistence
+    # intervals with zero length and can be skipped.
+
+    # This implementation of this optimization only works if (co)boundary simplices are
+    # returned in the correct order and if the birth times of σ and τ are the same.
+    emergent_check = emergent_pairs(matrix.filtration)
+    for cofacet in coboundary(matrix, column_simplex)
+        if emergent_check && birth(cofacet) == birth(column_simplex)
+            emergent_check = false
+            if !haskey(matrix.reduced, cofacet)
+                return cofacet_element(matrix)(cofacet)
+            end
+        end
+        nonheap_push!(matrix.chain, cofacet)
+    end
+    if isempty(matrix.chain)
+        return nothing
+    else
+        repair!(matrix.chain)
+        return pop!(matrix.chain)
+    end
+end
+
+function add!(matrix::CoboundaryMatrix, column, pivot)
+    factor = -coefficient(pivot)
+    for element in column
+        for cofacet in coboundary(matrix, simplex(element))
+            simplex(pivot) == cofacet && continue
+            push!(
+                matrix.chain,
+                cofacet_element(matrix)(cofacet, coefficient(element) * factor)
+            )
+        end
+    end
+    return matrix
+end
+
+function reduce_column!(matrix::CoboundaryMatrix, column_simplex)
+    pivot = initialize_coboundary!(matrix, column_simplex)
+
+    while !isnothing(pivot)
+        column = matrix.reduced[pivot]
[…]
