+++ src/Ripserer.jl
@@ -42,9 +42,12 @@ include("simplex.jl")
 include("abstractfiltration.jl")
 include("chainelement.jl")
 
-include("zerodimensional.jl")
-include("reductionmatrix.jl")
-include("main.jl")
+include("computation/utils.jl")
+include("computation/zerodimensional.jl")
+include("computation/reducedmatrix.jl")
+include("computation/workingchain.jl")
+include("computation/cohomology.jl")
+include("computation/ripserer.jl")
 
 include("filtrations/rips.jl")
 include("filtrations/cubical.jl")
