### 2020-09-17 a1ba65c
split out homology, implement less implicit algorithm
