+++ src/computation/workingchain.jl
@@ -54,265 +54,9 @@ function nonheap_push!(column::WorkingChain{E}, element::E) where E
 end
 
 repair!(column::WorkingChain) = heapify!(column.heap, column.ordering)
+
+function Base.sort!(column::WorkingChain; kwargs...)
+    return sort!(column.heap; order=column.ordering, kwargs...)
+end
+
 Base.first(column::WorkingChain) = first(column.heap)
-
-function move!(column::WorkingChain{E}) where E
-    dst = E[]
-    while (pivot = pop!(column)) ≠ nothing
-        push!(dst, pivot)
-    end
-    return dst
-end
-
-# ======================================================================================== #
-# TODO: the following code is pretty messy. There must be a way to handle all this with a
-# bit less type magic.
-struct ReductionMatrix{
-    Cohomology, T, Filtration, Simplex, SimplexElem, Cofacet, CofacetElem, O<:Base.Ordering
-}
-    filtration::Filtration
-    reduced::ReducedMatrix{Cofacet, SimplexElem, O}
-    working_coboundary::WorkingChain{CofacetElem, O}
-    columns_to_reduce::Vector{Simplex}
-    columns_to_skip::Vector{Simplex}
-end
-
-function ReductionMatrix{C, T}(
-    filtration::Filtration, columns_to_reduce, columns_to_skip
-) where {C, T, Filtration}
-
-    Simplex = eltype(columns_to_reduce)
-    cofacet_dim = dim(Simplex) + (C ? 1 : -1)
-    Cofacet = simplex_type(filtration, cofacet_dim)
-    ordering = C ? Base.Order.Forward : Base.Order.Reverse
-    O = typeof(ordering)
-    SimplexElem = chain_element_type(Simplex, T)
-    CofacetElem = chain_element_type(Cofacet, T)
-
-    reduced = ReducedMatrix{Cofacet, SimplexElem}(ordering)
-    sizehint!(reduced, length(columns_to_reduce))
-    working_coboundary = WorkingChain{CofacetElem}(ordering)
-
-    return ReductionMatrix{C, T, Filtration, Simplex, SimplexElem, Cofacet, CofacetElem, O}(
-        filtration,
-        reduced,
-        working_coboundary,
-        columns_to_reduce,
-        columns_to_skip
-    )
-end
-
-"""
-    coboundary(matrix, simplex)
-
-Iterate over the (co)boundary of the `simplex`. Chooses between the boundary and the
-coboundary based on the matrix type.
-"""
-function coboundary(matrix::ReductionMatrix{true}, simplex::AbstractSimplex)
-    return coboundary(matrix.filtration, simplex)
-end
-function coboundary(matrix::ReductionMatrix{false}, simplex::AbstractSimplex)
-    return boundary(matrix.filtration, simplex)
-end
-
-is_cohomology(::ReductionMatrix{C}) where C = C
-field_type(::ReductionMatrix{<:Any, F}) where F = F
-simplex_type(::ReductionMatrix{<:Any, <:Any, <:Any, S}) where S = S
-simplex_element(::ReductionMatrix{<:Any, <:Any, <:Any, <:Any, E}) where E = E
-cofacet_element(::ReductionMatrix{<:Any, <:Any, <:Any, <:Any, <:Any, <:Any, E}) where E = E
-dim(::ReductionMatrix{true, <:Any, <:Any, S}) where S = dim(S)
-dim(::ReductionMatrix{false, <:Any, <:Any, S}) where S = dim(S) - 1
-
-function initialize_coboundary!(matrix::ReductionMatrix, column_simplex)
-    empty!(matrix.working_coboundary)
-    # Emergent pairs: we are looking for pairs of simplices (σ, τ) where σ is the youngest
[…]
