### 2020-09-18 96f3a93
merge homology and cohomology, make distinction between implicit and explicit algorithm
