+++ src/Ripserer.jl
@@ -46,8 +46,7 @@ include("computation/utils.jl")
 include("computation/zerodimensional.jl")
 include("computation/workingchain.jl")
 include("computation/reducedmatrix.jl")
-include("computation/cohomology.jl")
-include("computation/homology.jl")
+include("computation/coboundarymatrices.jl")
 include("computation/ripserer.jl")
 
 include("filtrations/rips.jl")
