### 2020-09-18 219a19c
get rid of `Val`s
