### 2020-09-19 cf60440
implement implicit and explicit algorithms for both algorithms
