+++ src/computation/coboundarymatrices.jl
@@ -1,6 +1,3 @@
-# Not needed: most of the trait type functions defined on matrices...
-# Except dim, chain_element_type
-
 function initialize_coboundary!(matrix, column)
     initialize_coboundary!(Val(is_cohomology(matrix)), matrix, column)
 end
@@ -13,7 +10,7 @@ function initialize_coboundary!(::Val{true}, matrix, column)
 
     # This implementation of this optimization only works if (co)boundary simplices are
     # returned in the correct order and if the birth times of σ and τ are the same.
-    emergent_check = emergent_pairs(matrix.filtration)
+    emergent_check = emergent_pairs(matrix.filtration) && is_implicit(matrix)
     for cofacet in coboundary(matrix, column)
         if emergent_check && birth(cofacet) == birth(column)
             emergent_check = false
@@ -65,7 +62,7 @@ end
 function add!(::Val{false}, matrix, column, pivot)
     factor = -coefficient(pivot)
     for element in column
-        simplex(pivot) == simplex(element) && continue
+        # The pivot is not stored in the column, so there is no need to check for it.
         push!(matrix.chain, element * factor)
     end
 end
@@ -79,9 +76,8 @@ function finalize!(::Val{true}, matrix, column, pivot)
     commit!(matrix.reduced, simplex(pivot), inv(coefficient(pivot)))
 end
 # Explicit version
-function finalize!(::Val{false}, matrix, _, pivot)
-    push!(matrix.chain, pivot)
-    record!(matrix.reduced, matrix.chain, -coefficient(pivot))
+function finalize!(::Val{false}, matrix, column, pivot)
+    record!(matrix.reduced, matrix.chain)
     commit!(matrix.reduced, simplex(pivot), inv(coefficient(pivot)))
 end
 
@@ -114,12 +110,29 @@ function interval(matrix, column, pivot, cutoff, reps)
     else
         birth_simplex, death_simplex = simplex(pivot), column
     end
-    if reps && isnothing(pivot)
-        representative = eltype(matrix.reduced)[]
-    elseif reps
-        representative = collect(matrix.reduced[pivot])
-    else
+    if !reps
         representative = nothing
+    elseif is_cohomology(matrix)
+        if isnothing(pivot)
+            representative = simplex_type(matrix.filtration, dim(matrix))[]
+        elseif is_implicit(matrix)
+            representative = collect(matrix.reduced[pivot])
+        else
+            elem_t = chain_element_type(simplex_type(matrix.filtration, dim(matrix)), field_type(matrix))
+            tmp_chain = WorkingChain{elem_t}(Base.Order.Forward)
+            for elem in matrix.reduced[pivot]
+                for facet in boundary(matrix.filtration, simplex(elem))
+                    push!(tmp_chain, elem_t(facet, coefficient(elem)))
+                end
+            end
+            representative = move!(tmp_chain)
+        end
+    else
+        if is_implicit(matrix)
+            representative = pushfirst!(move!(matrix.chain), birth_simplex)
+        else
+            representative = pushfirst!(collect(matrix.reduced[pivot]), birth_simplex)
+        end
     end
     return interval(
[…]
