+++ src/computation/homology.jl
@@ -1,171 +1,0 @@
-struct BoundaryMatrix{T, F, S, R, C}
-    filtration::F
-    reduced::R
-    chain::C
-    columns_to_reduce::Vector{S}
-end
-
-function BoundaryMatrix(::Type{T}, filtration, columns_to_reduce, precomputed=()) where T
-    #TODO: move precomputed to reduced
-    Simplex = typeof(first(columns_to_reduce))
-    Facet = simplex_type(filtration, dim(Simplex) - 1)
-    ordering = Base.Order.Reverse
-    SimplexElem = chain_element_type(Simplex, T)
-    FacetElem = chain_element_type(Facet, T)
-
-    columns = Simplex[]
-    foreach(columns_to_reduce) do c
-        push!(columns, c)
-    end
-
-    reduced = ReducedMatrix{Facet, FacetElem}(ordering)
-    sizehint!(reduced, length(columns) + length(precomputed))
-    chain = WorkingChain{FacetElem}(ordering)
-
-    return BoundaryMatrix{T, typeof(filtration), Simplex, typeof(reduced), typeof(chain)}(
-        filtration, reduced, chain, columns
-    )
-end
-
-function boundary(matrix::BoundaryMatrix, simplex::AbstractSimplex)
-    return boundary(matrix.filtration, simplex)
-end
-
-field_type(::BoundaryMatrix{T}) where T = T
-simplex_type(::BoundaryMatrix{<:Any, <:Any, S}) where S = S
-simplex_element(::BoundaryMatrix{T, <:Any, S}) where {T, S} = chain_element_type(S, T)
-dim(bm::BoundaryMatrix) = dim(simplex_type(bm)) - 1
-facet_type(bm::BoundaryMatrix{<:Any, F}) where F = simplex_type(F, dim(bm) - 1)
-facet_element(bm::BoundaryMatrix{T}) where {T} = chain_element_type(facet_type(bm), T)
-
-function initialize_boundary!(matrix::BoundaryMatrix, column_simplex)
-    empty!(matrix.chain)
-    for facet in boundary(matrix, column_simplex)
-        nonheap_push!(matrix.chain, facet)
-    end
-    if isempty(matrix.chain)
-        return nothing
-    else
-        sort!(matrix.chain, alg=InsertionSort)
-        return pop!(matrix.chain)
-    end
-end
-
-function add!(matrix::BoundaryMatrix, column, pivot)
-    factor = -coefficient(pivot)
-    for element in column
-        simplex(pivot) == simplex(element) && continue
-        push!(matrix.chain, element * factor)
-    end
-    return matrix
-end
-
-function reduce_column!(matrix::BoundaryMatrix, column_simplex)
-    pivot = initialize_boundary!(matrix, column_simplex)
-
-    while !isnothing(pivot)
-        column = matrix.reduced[pivot]
-        isempty(column) && break
-
-        add!(matrix, column, pivot)
-        #record!(matrix.reduced, column, -coefficient(pivot))
-        pivot = pop!(matrix.chain)
-    end
-
-    if isnothing(pivot)
-        discard!(matrix.reduced)
-    else
-        push!(matrix.chain, pivot)
-        record!(matrix.reduced, matrix.chain, -coefficient(pivot))
-        commit!(matrix.reduced, simplex(pivot), inv(coefficient(pivot)))
-    end
-
-    return pivot
-end
-
-function add_interval!(intervals, matrix::BoundaryMatrix, birth_sx, death_sx, cutoff)
-    if !isnothing(birth_sx)
-        birth_time = Float64(birth(birth_sx))
-        death_time = Float64(birth(death_sx))
-        if death_time - birth_time > cutoff
-            meta = (;
-                    birth_simplex=birth_sx,
[…]

[…]
