### 2020-09-20 4070d70
minor perf tweaks, sort by persistence
