+++ CHANGELOG.md
@@ -1,1 +1,5 @@
 # v0.15.0
+
+* Results are now sorted by persistence instead of birth time.
+* Homology is now computed with the `alg=:homology` keyword argument.
+*
