### 2020-09-20 f704aed
test new algorithms, bugfix
