+++ src/computation/coboundarymatrices.jl
@@ -1,258 +1,3 @@
-function initialize_coboundary!(matrix, column)
-    initialize_coboundary!(Val(is_cohomology(matrix)), matrix, column)
-end
-# Cohomology version
-function initialize_coboundary!(::Val{true}, matrix, column)
-    empty!(matrix.chain)
-    # Emergent pairs: we are looking for pairs of simplices (σ, τ) where σ is the youngest
-    # facet of τ and τ is the oldest cofacet of σ. These pairs give birth to persistence
-    # intervals with zero length and can be skipped.
-
-    # This implementation of this optimization only works if (co)boundary simplices are
-    # returned in the correct order and if the birth times of σ and τ are the same.
-    emergent_check = emergent_pairs(matrix.filtration) && is_implicit(matrix)
-    for cofacet in coboundary(matrix, column)
-        if emergent_check && birth(cofacet) == birth(column)
-            emergent_check = false
-            if !haskey(matrix.reduced, cofacet)
-                return chain_element_type(matrix)(cofacet)
-            end
-        end
-        nonheap_push!(matrix.chain, cofacet)
-    end
-    if isempty(matrix.chain)
-        return nothing
-    else
-        repair!(matrix.chain)
-        return pop!(matrix.chain)
-    end
-end
-# Homology version
-function initialize_coboundary!(::Val{false}, matrix, column)
-    empty!(matrix.chain)
-    for facet in coboundary(matrix, column)
-        nonheap_push!(matrix.chain, facet)
-    end
-    if isempty(matrix.chain)
-        return nothing
-    else
-        repair!(matrix.chain)
-        return pop!(matrix.chain)
-    end
-end
-
-function add!(matrix, column, pivot)
-    add!(Val(is_implicit(matrix)), matrix, column, pivot)
-end
-# Implicit version
-function add!(::Val{true}, matrix, column, pivot)
-    factor = -coefficient(pivot)
-    for element in column
-        for cofacet in coboundary(matrix, simplex(element))
-            simplex(pivot) == cofacet && continue
-            push!(
-                matrix.chain,
-                chain_element_type(matrix)(cofacet, coefficient(element) * factor),
-            )
-        end
-    end
-    record!(matrix.reduced, column, factor)
-end
-# Explicit version
-function add!(::Val{false}, matrix, column, pivot)
-    factor = -coefficient(pivot)
-    for element in column
-        # The pivot is not stored in the column, so there is no need to check for it.
-        push!(matrix.chain, element * factor)
-    end
-end
-
-function finalize!(matrix, column, pivot)
-    finalize!(Val(is_implicit(matrix)), matrix, column, pivot)
-end
-# Implicit version
-function finalize!(::Val{true}, matrix, column, pivot)
-    record!(matrix.reduced, column)
-    commit!(matrix.reduced, simplex(pivot), inv(coefficient(pivot)))
-end
-# Explicit version
-function finalize!(::Val{false}, matrix, column, pivot)
-    record!(matrix.reduced, matrix.chain)
-    commit!(matrix.reduced, simplex(pivot), inv(coefficient(pivot)))
-end
-
-function reduce_column!(matrix, column_to_reduce)
-    pivot = initialize_coboundary!(matrix, column_to_reduce)
-
-    while !isnothing(pivot)
-        column = matrix.reduced[pivot]
-        isempty(column) && break
-
-        add!(matrix, column, pivot)
-        pivot = pop!(matrix.chain)
-    end
-    if isnothing(pivot)
-        discard!(matrix.reduced)
-    else
-        finalize!(matrix, column_to_reduce, pivot)
-    end
-
-    return pivot
-end
-
-function interval(matrix, column, pivot, cutoff, reps)
[…]
