+++ src/computation/ripserer.jl
@@ -60,7 +60,7 @@ function ripserer(
     field_type=Mod{modulus},
     progress=false,
     alg=:cohomology,
-    implicit=alg == :cohomology,
+    implicit=alg != :homology,
 )
     start_time = time_ns()
     index_overflow_check(filtration, field_type, dim_max)
@@ -93,7 +93,6 @@ function _ripserer(
     return result
 end
 
-# Homology: reps is on by default for dimensions above 0.
 function _ripserer(
     ::Val{:homology}, filtration, cutoff, progress, field_type, dim_max, reps, implicit
 )
@@ -118,11 +117,29 @@ function _ripserer(
     return result
 end
 
-# Assisted homology: reps is on by default for dimensions above 0.
-# TODO
 function _ripserer(
     ::Val{:assisted}, filtration, cutoff, progress, field_type, dim_max, reps, implicit
 )
+    result = PersistenceDiagram[]
+    zeroth, to_reduce, to_skip = zeroth_intervals(
+        filtration, cutoff, progress, field_type, Val(reps)
+    )
+    push!(result, zeroth)
+    if dim_max > 0
+        comatrix = CoboundaryMatrix{true}(field_type, filtration, to_reduce, to_skip)
+        for dim in 1:dim_max
+            columns = compute_death_simplices!(comatrix, progress)
+            if isempty(columns)
+                return result
+            end
+            matrix = BoundaryMatrix{implicit}(field_type, filtration, columns)
+            push!(result, compute_intervals!(matrix, cutoff, progress, true))
+            if dim < dim_max
+                comatrix = next_matrix(comatrix, progress)
+            end
+        end
+    end
+    return result
 end
 
 function _ripserer(::Val{A}, args...) where A
