### 2020-09-20 dd32691
find threshold with assisted homology, sort zeroth intervals
