### 2020-09-22 b6d0ad7
clean up interval creation
