+++ src/computation/coboundarymatrices.jl
@@ -1,3 +1,9 @@
+"""
+    CoboundaryMatrix{I}
+
+This `struct` is used to compute cohomology. The `I` parameter sets whether the implicit
+algoritm is used or not.
+"""
 struct CoboundaryMatrix{I, T, F, S, R, C}
     filtration::F
     reduced::R
@@ -29,7 +35,6 @@ function CoboundaryMatrix{I}(
         filtration, reduced, chain, columns_to_reduce, columns_to_skip
     )
 end
-
 
 field_type(::CoboundaryMatrix{<:Any, T}) where T = T
 dim(cm::CoboundaryMatrix{<:Any, <:Any, <:Any, S}) where S = dim(S)
@@ -72,9 +77,12 @@ function next_matrix(matrix::CoboundaryMatrix{I}, progress) where I
     )
 end
 
-###
-###
-###
+"""
+    BoundaryMatrix{I}
+
+This `struct` is used to compute homology. The `I` parameter sets whether the implicit
+algoritm is used or not.
+"""
 struct BoundaryMatrix{I, T, F, S, R, C}
     filtration::F
     reduced::R
@@ -118,18 +126,20 @@ end
 is_implicit(::BoundaryMatrix{I}) where I = I
 is_cohomology(::BoundaryMatrix) = false
 
+# The naming here is not ideal...
 function coboundary(matrix::BoundaryMatrix, simplex::AbstractSimplex)
     return boundary(matrix.filtration, simplex)
 end
 
-###
-###
-###
-function initialize_coboundary!(matrix, column)
-    initialize_coboundary!(Val(is_cohomology(matrix)), matrix, column)
-end
-# Cohomology version
-function initialize_coboundary!(::Val{true}, matrix, column)
+"""
+    initialize_coboundary!(matrix, column)
+
+Initialize the column indexed by `column` by adding its (co)boundary to `matrix.chain`. This
+is where the emergent pairs optimization gets triggered for implicit versions of the
+algorithm.
+
+"""
+function initialize_coboundary!(#=::Val{true},=# matrix, column)
     empty!(matrix.chain)
     # Emergent pairs: we are looking for pairs of simplices (σ, τ) where σ is the youngest
     # facet of τ and τ is the oldest cofacet of σ. These pairs give birth to persistence
@@ -154,20 +164,14 @@ function initialize_coboundary!(::Val{true}, matrix, column)
         return pop!(matrix.chain)
     end
 end
-# Homology version
-function initialize_coboundary!(::Val{false}, matrix, column)
-    empty!(matrix.chain)
-    for facet in coboundary(matrix, column)
-        nonheap_push!(matrix.chain, facet)
-    end
-    if isempty(matrix.chain)
-        return nothing
-    else
-        repair!(matrix.chain)
-        return pop!(matrix.chain)
-    end
-end
 
+"""
+    add!(matrix, column, pivot)
+
+Add already reduced column indexed by `column` multiplied by `-coefficient(pivot)` to
+`matrix.chain`.
+
+"""
 function add!(matrix, column, pivot)
     add!(Val(is_implicit(matrix)), matrix, column, pivot)
 end
@@ -194,6 +198,11 @@ function add!(::Val{false}, matrix, column, pivot)
     end
 end
 
+"""
+    finalize!(matrix, column, pivot)
+
+After reduction is done, finalize the column by adding it to the reduced matrix.
+"""
 function finalize!(matrix, column, pivot)
     finalize!(Val(is_implicit(matrix)), matrix, column, pivot)
 end
@@ -208,6 +217,13 @@ function finalize!(::Val{false}, matrix, column, pivot)
[…]

[…]
