### 2020-09-22 2f268b7
update `ripserer` docstring
