### 2020-09-22 7d2335e
progress and docstring tweaks
