### 2020-09-22 f222c5d
remove unused functionality, two more tests
