### 2020-09-23 211905a
add infinite interval support for :assisted
