### 2020-09-23 4abca53
Taking homology seriously (#103)
* computation code overhaul

* new homology algorithms
