+++ test/filtrations/rips.jl
@@ -341,7 +341,7 @@ end
             end
             # 0-dimensional persistence should find values of minima and maxima of our data.
             d0 = ripserer(dists; dim_max=0)[1]
-            @test d0 == [(0, Inf), (1, 4), (2, 3)]
+            @test d0 == [(2, 3), (1, 4), (0, Inf)]
         end
 
         @testset "with Rips" begin
@@ -357,37 +357,53 @@ end
                 end
             end
             d0 = ripserer(dists; dim_max=0)[1]
-            @test d0 == [(0, Inf), (1, 4), (2, 3)]
+            @test d0 == [(2, 3), (1, 4), (0, Inf)]
         end
     end
 
-    @testset "Persistent homology" begin
-        @testset "Produces the same diagram as cohomology" begin
-            res_hom = ripserer(cycle, cohomology=false, dim_max=3)
-            res_coh = ripserer(cycle, cohomology=true, dim_max=3)
-
-            @test res_hom == res_coh
-        end
-        @testset "Has the same cirical simplices as cohomology" begin
-            # Add some noise because critical simplices might be different if values are
-            # exactly the same.
-            cyc = cycle .+ 0.01 .* rand_dist_matrix(18)
-            res_hom = ripserer(cyc, cohomology=false, dim_max=3, reps=true)
-            res_coh = ripserer(cyc, cohomology=true, dim_max=3, reps=true)
-
-            for i in 1:4
-                @test birth_simplex.(res_hom[i]) == birth_simplex.(res_coh[i])
-                @test death_simplex.(res_hom[i]) == death_simplex.(res_coh[i])
+    @testset "Homology and explicit cohomology" begin
+        for m in (2, 17)
+            @testset "Same for projective plane with modulus=$m" begin
+                r = Rips(projective_plane)
+                _, hom_imp1, hom_imp2 = ripserer(
+                    r; alg=:homology, implicit=true, dim_max=2, modulus=m
+                )
+                _, hom_exp1, hom_exp2 = ripserer(
+                    r; alg=:homology, implicit=false, dim_max=2, modulus=m
+                )
+                _, hom_ass1, hom_ass2 = ripserer(
+                    r; alg=:assisted, dim_max=2, modulus=m
+                )
+                _, coh_imp1, coh_imp2 = ripserer(
+                    r ; implicit=true, reps=true, dim_max=2, modulus=m
+                )
+                _, coh_exp1, coh_exp2 = ripserer(
+                    r; implicit=true, reps=true, dim_max=2, modulus=m
+                )
+                @test hom_imp1 == hom_exp1 == hom_ass1 == coh_imp1 == coh_exp1
+                @test hom_imp2 == hom_exp2 == hom_ass2 == coh_imp2 == coh_exp2
+                @test representative.(hom_imp1) == representative.(hom_exp1)
+                @test representative.(hom_imp1) == representative.(hom_ass1)
+                @test representative.(hom_imp2) == representative.(hom_exp2)
+                @test representative.(hom_imp2) == representative.(hom_ass2)
+                @test representative.(coh_imp1) == representative.(coh_exp1)
+                @test representative.(coh_imp2) == representative.(coh_exp2)
             end
         end
         @testset "Representative cycle" begin
-            res_hom = ripserer(cycle, cohomology=false, reps=true, dim_max=3)
+            res_hom = ripserer(cycle, alg=:homology, reps=true, dim_max=3)
             @test vertices.(simplex.(representative(res_hom[2][1]))) == sort!(vcat(
                 [SVector(i+1, i) for i in 1:17], [SVector(18, 1)]
             ))
         end
[…]
