+++ src/computation/reducedmatrix.jl
@@ -1,0 +1,152 @@
+"""
+    ReducedMatrix{S<:AbstractSimplex, E<:AbstractChainElement}
+
+Representation of the reduced part of the matrix. Is indexed by `S`, indexing iterates
+values of type `E`. `is_homology` controls the ordering of values.
+
+Changes to the matrix are recorded in a buffer with `record!`. Once a column is ready to be
+added, use `commit!` to move the changes from the buffer to the matrix and create a
+column. Changes can also be discarded with `discard!`.
+"""
+struct ReducedMatrix{S<:AbstractSimplex, E<:AbstractChainElement, O<:Base.Ordering}
+    column_index::Dict{S, Int}
+    indices::Vector{Int}
+    values::Vector{E}
+    buffer::Vector{E}
+    ordering::O
+
+    function ReducedMatrix{S, E}(ordering::O) where {S, E, O}
+        return new{S, E, O}(Dict{S, Int}(), Int[1], E[], E[], ordering)
+    end
+end
+
+function Base.sizehint!(matrix::ReducedMatrix, size)
+    sizehint!(matrix.column_index, size)
+    sizehint!(matrix.indices, size + 1)
+    sizehint!(matrix.values, size)
+    return matrix
+end
+
+"""
+    record!(matrix::ReducedMatrix, element)
+    record!(matrix::ReducedMatrix, elements, factor)
+    record!(matrix::ReducedMatrix, chain, factor)
+
+Record the operation that was performed in the buffer. If the second argument is a
+`WorkingChain`, it is emptied into the buffer.
+"""
+function record!(matrix::ReducedMatrix{<:Any, E}, simplex::AbstractSimplex) where E
+    return push!(matrix.buffer, E(simplex))
+end
+
+function record!(matrix::ReducedMatrix, elements, factor)
+    i = length(matrix.buffer)
+    resize!(matrix.buffer, length(matrix.buffer) + length(elements))
+    @inbounds for element in elements
+        i += 1
+        matrix.buffer[i] = element * factor
+    end
+    return elements
+end
+
+function record!(matrix::ReducedMatrix, chain::WorkingChain)
+    while (element = pop!(chain)) ≢ nothing
+        push!(matrix.buffer, element)
+    end
+    return chain
+end
+
+"""
+    commit!(matrix::ReducedMatrix, simplex, factor)
+
+Commit the changes that were `record!`ed, creating a new column indexed by `simplex`. This
+sorts the buffer and adds duplicates together. All entries are multiplied by `factor`.
+"""
+function commit!(matrix::ReducedMatrix, simplex, factor)
+    @assert sign(simplex) == 1
+    isempty(matrix.buffer) && return matrix
+
+    sort!(matrix.buffer, alg=QuickSort, order=matrix.ordering)
+    i = 0
+    @inbounds prev = matrix.buffer[1]
+    @inbounds for j in 2:length(matrix.buffer)
+        current = matrix.buffer[j]
+        if current == prev
+            prev += current
+        else
+            if !iszero(prev)
+                i += 1
+                matrix.buffer[i] = prev * factor
+            end
+            prev = current
+        end
+    end
+    if !iszero(prev)
+        i += 1
+        @inbounds matrix.buffer[i] = prev * factor
+    end
+    if i > 0
+        resize!(matrix.buffer, i)
+        append!(matrix.values, matrix.buffer)
+        push!(matrix.indices, matrix.indices[end] + i)
+        matrix.column_index[simplex] = length(matrix.indices) - 1
+    end
+
+    empty!(matrix.buffer)
+    return matrix
+end
+
[…]
