### 2020-09-24 8aeafd3
move radius and friends to separate file
