+++ src/filtrations/rips.jl
@@ -73,9 +73,9 @@ end
         n = length(vertices)
         diameter = typemin(T)
         for i in 1:n, j in i+1:n
-            d = adj[vertices[j], vertices[i]]
-            (iszero(d) || d > threshold(rips)) && return nothing
-            diameter = ifelse(d > diameter, d, diameter)
+            e = adj[vertices[j], vertices[i]]
+            (iszero(e) || e > threshold(rips)) && return nothing
+            diameter = ifelse(e > diameter, e, diameter)
         end
         return S(I(sign) * index(vertices), diameter)
     end
@@ -95,9 +95,9 @@ end
         v == new_vertex && continue
         # Even though this looks like a tight loop, v changes way more often than us, so
         # this is the faster order of indexing by new_vertex and v.
-        d = adj[new_vertex, v]
-        (iszero(d) || d > threshold(rips)) && return nothing
-        diameter = ifelse(d > diameter, d, diameter)
+        e = adj[new_vertex, v]
+        (iszero(e) || e > threshold(rips)) && return nothing
+        diameter = ifelse(e > diameter, e, diameter)
     end
     return S(I(sign) * index(cofacet_vertices), diameter)
 end
@@ -128,61 +128,6 @@ function coboundary(
     else
         return Coboundary{A}(rips, sx)
     end
-end
-
-"""
-    to_matrix(points)
-
-Convert collection of d-dimensional points to d×n matrix
-"""
-function to_matrix(points)
-    dim = length(points[1])
-    T = eltype(points[1])
-    n = length(points)
-    result = zeros(T, (dim, n))
-
-    for (i, p) in enumerate(points)
-        length(p) == dim || throw(ArgumentError("points must have the same length"))
-        result[:, i] .= p
-    end
-
-    return result
-end
-
-"""
-    distances(points, metric=Euclidean(1e-12))
-
-Return distance matrix calculated from `points` with `metric`.
-"""
-function distances(points, metric=Euclidean(1e-12))
-    points_mat = to_matrix(points)
-    dists = pairwise(metric, points_mat, dims=2)
-    return dists
-end
-
-"""
-    radius(dists)
-    radius(points[, metric=Euclidean(1e-12)])
-
-Calculate the radius of the space. This is used for default `thresholds`.
-"""
-function radius(dists::AbstractMatrix)
-    return minimum(maximum(abs, dists[:, i]) for i in 1:size(dists, 1))
-end
-function radius(dists::SparseMatrixCSC)
-    return maximum(dists)
-end
-function radius(points, metric=Euclidean(1e-12))
-    radius = Inf
-    for p in points
-        p_max = 0.0
-        for q in points
-            p == q && continue
-            p_max = max(p_max, metric(SVector(p), SVector(q)))
-        end
-        radius = min(p_max, radius)
-    end
-    return radius
 end
 
 function _check_distance_matrix(dist::SparseMatrixCSC)
