### 2020-09-24 77c6cd8
fix benchmarking code
