+++ src/computation/ripserer.jl
@@ -91,7 +91,7 @@ function ripserer(
         Val(alg), filtration, cutoff, progress, field_type, dim_max, reps, implicit
     )
     elapsed = round((time_ns() - start_time) / 1e9, digits=3)
-    prog_println(progress, "Done. Took ", elapsed, " seconds.")
+    prog_println(progress, "Done. Time: ", ProgressMeter.durationstring(elapsed))
 
     return result
 end
