+++ src/Ripserer.jl
@@ -19,7 +19,6 @@ using Distances
 using IterTools
 using MiniQhull
 using PersistenceDiagrams
-using Printf
 using ProgressMeter
 using RecipesBase
 using StaticArrays
