### 2020-09-25 c6e755f
Organization before v0.15 (#104)
* move radius and friends to separate file

* rename assisted to involuted, fix tests

* fix benchmarking code

* fix :assisted occurence

* make printing prettier

* fix test, remove Printf

* bump version

* fix travis badge
