### 2020-09-25 4ef7332
rename dist to distance_matrix
