### 2020-09-25 7c183e1
Representative cycles from cocycles (#101)
* initial representative cycle implementation

* improved implementation

* use A* instead of Floyd Warshall
* add `dist` for all filtrations
* tests

* add underscores to cubical local functions

* add tests

* adapt to new master

* fix a_star heuristic

* rename dist to distance_matrix

* add option to change distance matrix
