+++ src/filtrations/cubical.jl
@@ -13,7 +13,7 @@
 # the birth times of cofaces.
 #
 # See https://link.springer.com/chapter/10.1007%2F978-3-642-23175-9_7 for more info.
-@generated function cubemap(input::Array{T, N}) where {T, N}
+@generated function _cubemap(input::Array{T, N}) where {T, N}
     quote
         result = similar(input, size(input) .* 2 .- 1)
         result .= typemin(T)
@@ -27,16 +27,16 @@
     end
 end
 
-one_hot(i, ::Val{N}) where N = CartesianIndex{N}(ntuple(isequal(i), Val(N)))
+_one_hot(i, ::Val{N}) where N = CartesianIndex{N}(ntuple(isequal(i), Val(N)))
 
 # Convenience functions from converting cubemap index to vertices and back.
-@inline function from_cubemap(root::CartesianIndex{K}, ::Val{N}) where {K, N}
+@inline function _from_cubemap(root::CartesianIndex{K}, ::Val{N}) where {K, N}
     2^count(iseven, Tuple(root)) == N || throw(ArgumentError("invalid N"))
     result = ntuple(_ -> root, Val(N))
     for (i, j) in enumerate(Tuple(root))
         if iseven(j)
             result = result .+ ntuple(Val(N)) do k
-                ifelse(isodd(k), 1, -1) * one_hot(i, Val(K))
+                ifelse(isodd(k), 1, -1) * _one_hot(i, Val(K))
             end
         end
         result = TupleTools.sort(result, by=Base.Fix2(getindex, i))
@@ -47,18 +47,23 @@ one_hot(i, ::Val{N}) where N = CartesianIndex{N}(ntuple(isequal(i), Val(N)))
     end
 end
 
-function to_cubemap(vertices::NTuple{N}) where N
-    floor(log2(N)) == log2(N) || throw(ArgumentError("number vertices is not 2^d"))
-    allunique(vertices) || throw(ArgumentError("duplicate vertices in $vertices"))
+function _to_cubemap(vertices::NTuple{N}) where N
     K = length(first(vertices))
-    for i in 1:K
-        l, h = extrema(getindex.(vertices, i))
-        h - l ≤ 1 || throw(ArgumentError("vertices not adjacent"))
-    end
     result = ntuple(Val(K)) do i
         sum(2 .* getindex.(vertices, i) .- 1) ÷ N
     end
     return CartesianIndex{K}(result)
+end
+
+function _is_valid(vertices::NTuple{N}) where N
+    floor(log2(N)) == log2(N) || return false
+    allunique(vertices) || return false
+    K = length(first(vertices))
+    for i in 1:K
+        l, h = extrema(getindex.(vertices, i))
+        h - l ≤ 1 || return false
+    end
+    return true
 end
 
 """
@@ -90,7 +95,9 @@ function Cube{D}(root::CartesianIndex{K}, birth::T) where {D, T, K}
     return Cube{D, T, K}(root, birth)
 end
 function Cube{D}(vertices, birth) where D
-    root = to_cubemap(tuple(CartesianIndex.(vertices)...))
+    vs = tuple(CartesianIndex.(vertices)...)
+    _is_valid(vs) || throw(ArgumentError("invalid vertices"))
+    root = _to_cubemap(vs)
     return Cube{D}(root, birth)
 end
 
@@ -101,7 +108,7 @@ Base.:-(cube::Cube) = cube
 Base.abs(cube::Cube) = cube
 
 @generated Base.length(::Type{<:Cube{D}}) where D = :($(2^D))
-vertices(cube::Cube{D}) where D = from_cubemap(index(cube), Val(length(cube)))
[…]
