+++ docs/src/benchmarks.md
@@ -1,9 +1,10 @@
 # Benchmarks
 
 The following tables show benchmarks that compare Ripserer's performance with
-[Ripser](https://github.com/Ripser/ripser) and [Cubical
-Ripser](https://github.com/CubicalRipser/). The benchmarking code and more info about the
-datasets is available [here](https://github.com/mtsch/RipsererBenchmarks.jl).
+[Ripser](https://github.com/Ripser/ripser), [Cubical
+Ripser](https://github.com/CubicalRipser/), and
+[Eirene.jl](https://github.com/Eetion/Eirene.jl). The benchmarking code and more info about
+the datasets are available [here](https://github.com/mtsch/RipsererBenchmarks.jl).
 
 All benchmarks were performed on a laptop with an Intel(R) Core(TM) i5-4200U CPU @ 1.60GHz
 with 8GB of RAM.
@@ -13,99 +14,88 @@ timing benchmarks and [Valgrind's Massif
 tool](https://www.valgrind.org/docs/manual/ms-manual.html) to measure peak heap sizes
 (i.e. total memory footprint).
 
-All benchmarks include I/O, so there might be some noise in the smaller benchmarks. For
-Ripserer, [CSV.jl](https://github.com/JuliaData/CSV.jl) was used to read the
-files. Ripserer's memory footprint benchmarks include the Julia runtime, which tends to use
-around 100MiB memory.
+The benchmarks were performed with Ripserer v0.15, `master` versions of Ripser (commit
+hash `286d369`) and Cubical Ripser (commit hashes `6edb9c5` for 2D and `a063dac` for 3D),
+and Eirene v1.3.5.
 
-The benchmarks were performed with Ripserer v0.14.3 and `master` versions of Ripser (commit
-hash `286d369`) and Cubical Ripser (commit hashes `6edb9c5` for 2D and `a063dac` for 3D).
+The timings show the minimum time taken among five runs of the benchmark.
 
-The timings show the minimum time taken for five repetitions of the benchmark.
+The heap sizes for Ripserer include the Julia runtime.
 
-## Vietoris-Rips Persistent Homology
+## Comparison with Ripser
 
 In this experiment, we performed benchmarks with the datasets presented in the [Ripser
 article](https://arxiv.org/abs/1908.02518). We only used the datasets that we were able to
-run with less than 8GB memory. `ripserer_s` denotes `ripserer` run with `sparse=true`. All
-datasets were parsed as `Float32` as that is what Ripser supports.
+run with less than 8GB memory. All datasets were parsed as `Float32` as that is what Ripser
+supports. The time it takes to parse a file is included for both Ripser and Ripserer.
 
-The timing results are in the table below. The ratio column shows the better ratio.
+### Dense results
 
-| dataset    | size  | `dim_max` | `threshold` | `ripserer` | `ripserer_s` | `ripser`  | ratio |
-| :--------- | :---- | :-------- | :---------- | :--------- | :----------- | :-------- | :---- |
-| `random16` | 50    | 2         |             | 8.1ms      | 9.3ms        | 10.2ms   | 0.80  |
-| `o3_1024`  | 1024  | 3         | 1.8         | 4.3s       | 3.1s         | 3.1s     | 1.01  |
[…]
