### 2020-09-30 79f772d
Resolve ambiguities with `convert` (#106)
* resolve ambiguities

* fix for 1.0
