### 2020-10-01 e48b0a5
improve representative section in docs
