+++ docs/src/examples/cocycles.jl
@@ -1,55 +1,169 @@
-# # Representative Cocycles
+# # Cohomology, Homology, and Representatives
 
-# In this example, we will demonstrate how to compute and visualize representative cocycles
-# of persistent homology generators.
+# In this section, we will show how Ripserer can be used to find critical simplices and
+# representative (co)cycles.
 
-# As always, we start by importing the relevant packages.
-using Ripserer
+# We start by loading some packages and generating some data.
+
+using LinearAlgebra
 using Plots
-gr(); nothing # hide
+using Ripserer
+using Random
+Random.seed!(1337)
+gr() # hide
+nothing # hide
 
-# We define a function that generates some data sampled from a curve.
-
-function curve(n)
-    [(sin(t)+t/10, cos(t)+t/10) for t in range(0, 2π, length=n)]
-end
-
-plot(curve(100), legend=false, title="Curve", aspect_ratio=1, xlab="x", ylab="y")
-
-# Then, we compute the persistent homology of that data with various numbers of points. We
-# plot the representative cocycles of each result and turn them into an animation. This data
-# set will always have at most one one-dimensional class. We invoke `plot(interval, data)`
-# to plot the representative cocycle of the class. The same plot could be created by
-# invoking `representative(interval)`. We use the `threshold_strict` argument to only plot
-# the simplices with diameter strictly lower than the death time of the interval.
-
-anim = @animate for i in vcat(3:100, 100:-1:3)
-    points = curve(i)
-    res = ripserer(points, reps=true)
-
-    plt1 = plot(title="1-dimensional Representative Cocycle",
-                xlab="x",
-                ylab="y",
-                legend=false,
-                ylim=(-0.8, 1.8),
-                aspect_ratio=1)
-    if length(res[2]) > 0
-        interval = res[2][1]
-        plot!(plt1, interval, points,
-              alpha=0.2,
-              linewidth=2,
-              threshold_strict=death(interval),
-              color=3)
+function annulus(n, r1=1, r2=2, offset=(0,0))
+    result = Tuple{Float64, Float64}[]
+    while length(result) < n
+        point = 2 * r2 * rand(2) .- r2
+        if r1 < norm(point) < r2
+            push!(result, (point[1] + offset[1], point[2] + offset[2]))
+        end
     end
-    scatter!(plt1, points, color=2)
-    plot(plt1, barcode(res), plot(res),
-         layout=@layout([a [b; c]]),
-         size=(800, 600))
+    return result
 end
-mp4(anim, "cocycles_anim.mp4") # hide
 
-# We can save the animation by running the following.
+data = annulus(300)
 
-# ```julia
-# mp4(anim, "cocycles_anim.mp4")
-# ```
+scatter(data, label="data", markersize=1)
+
+# Let's start by taking a look at the persistence diagram.
+
+diagram = ripserer(data)
+plot(diagram)
+
+# The diagram tells us that there is a persistent hole in the data, but tells us nothing
+# about the location of the hole. Ripserer provides several methods to locate it. We'll
+# start with the simplest.
+
+# ## Critical simplices
+
+# The first option is to find the death simplex of the interval.
+
+# We start by extracting the interval in question. Keep in mind that the diagrams are sorted
+# by persistence, so the last element will always be the most persistent.
+
+most_persistent = diagram[2][end]
+
+# Notice that the interval has two simplices attached to it, the birth simplex and the death
+# simplex. We can extract them with `birth_simplex` and `death_simplex` respectively.
+
+death_sx = death_simplex(most_persistent)
+
+# A simplex acts just like an array of indices, so it can be used to index into the data.
+
+data[death_sx]
+
[…]
