### 2020-10-02 e490717
Improve representatives docs (#107)
* improve representative section in docs

* update
