### 2020-10-02 88b1f5a
Add benchmark that favors Eirene
