### 2020-10-02 a00ab95
don't report fractions of miliseconds
