### 2020-10-02 02b5f5c
Add benchmark that favors Eirene (#108)
* Add benchmark that favors Eirene

* don't report fractions of miliseconds
