### 2020-10-02 #2614 @VivianePons
Hi @sauln I haven't heard of you in a while and nothing has been udptaed on the review, are you still on board?
