### 2020-10-06 f0efb51
Compute cocycles for infinite intervals
