### 2020-10-06 e90f55f
Revert progress meter dt in zerodimensional, bump version (#110)
* Revert progress meter dt in zerodimensional, bump version

* update changelog
