### 2020-10-09 #2614 @sauln
@VivianePons, I have completed my review.  As @juliohm expressed, this project was a pleasure to review. I am only sorry it took me so long to complete it.  The documentation is best in class, the library interface is remarkably clean and intuitive, and the testing appears more exhaustive than any other TDA library I've seen.

I am happy to accept this submission to JOSS 🥇
