### 2020-10-12 #2614 @VivianePons
@mtsch I see you give many arXiv references in the bibliography. the arXiv number is always given but the actual link is present only in some ref. Could you put the link on all of them, that makes it easier for the reader to follow the ref!

Once this is done, I need you to make an archive (on Zenodo, or figshare, or other) and give the doi.

Thank you very much!
