### 2020-10-12 8db9e34
JOSS: Add missing links to references (#112)
