+++ joss/paper.bib
@@ -57,7 +57,8 @@
   title={Cubical Ripser: Software for computing persistent homology of image and volume data},
   author={Kaji, Shizuo and Sudo, Takeki and Ahara, Kazushi},
   journal={arXiv preprint arXiv:2005.12692},
-  year={2020}
+  year={2020},
+  url={https://arxiv.org/pdf/2005.12692.pdf}
 }
 @article{lee2017quantifying,
   title={Quantifying similarity of pore-geometry in nanoporous materials},
@@ -100,7 +101,8 @@
   title={GPU-Accelerated Computation of Vietoris-Rips Persistence Barcodes},
   author={Zhang, Simon and Xiao, Mengbai and Wang, Hao},
   journal={arXiv preprint arXiv:2003.07989},
-  year={2020}
+  year={2020},
+  url={https://arxiv.org/abs/2003.07989}
 }
 @article{bezanson2017julia,
   title={Julia: A fresh approach to numerical computing},
@@ -118,14 +120,16 @@
   title={Matroid filtrations and computational persistent homology},
   author={Henselman, Gregory and Ghrist, Robert},
   journal={arXiv preprint arXiv:1606.00199},
-  year={2016}
+  year={2016},
+  url={https://arxiv.org/abs/1606.00199}
 }
 
 @article{brehm2018sparips,
   title={Sparips},
   author={Brehm, Bernhard and Hardering, Hanne},
   journal={arXiv preprint arXiv:1807.09982},
-  year={2018}
+  year={2018},
+  url={https://arxiv.org/abs/1807.09982}
 }
 @article{lutgehetmann2020computing,
   title={Computing persistent homology of directed flag complexes},
