+++ src/Ripserer.jl
@@ -29,9 +29,6 @@ using TupleTools
 # piracy involved here.
 import LightGraphs: vertices, edges, nv, adjacency_matrix
 
-import MLJModelInterface
-const MMI = MLJModelInterface
-
 # Reexporting basics makes Ripserer usable without having to import another package.
 import PersistenceDiagrams: birth, threshold, dim
 export birth, death, persistence, representative, birth_simplex, death_simplex, barcode
@@ -72,8 +69,14 @@ include("filtrations/custom.jl")
 include("filtrations/alpha.jl")
 
 include("cycles.jl")
-
-include("mlj.jl")
 include("simplexrecipes.jl")
+
+# WIP
+include("mlj.jl")
+module MLJRipserer
+using ..Ripserer: RipsPersistentHomology, AlphaPersistentHomology, CubicalPersistentHomology
+export RipsPersistentHomology, AlphaPersistentHomology, CubicalPersistentHomology
+end
+export MLJRipserer
 
 end
