### 2020-10-17 bdc1872
Fix type instability in zeroth interval generation
