### 2020-10-17 b5fcbf3
allow failures on nightly, another small performance tweak in interval construction
