### 2020-10-17 8d23372
rework fix to work with cubical homology
