### 2020-10-17 b03f274
Fix type instability in zeroth interval generation (#116)
* Fix type instability in zeroth interval generation

* bump version

* allow failures on nightly, another small performance tweak in interval construction

* rework fix to work with cubical homology
