+++ src/filtrations/cubical.jl
@@ -13,7 +13,7 @@
 # the birth times of cofaces.
 #
 # See https://link.springer.com/chapter/10.1007%2F978-3-642-23175-9_7 for more info.
-@generated function _cubemap(input::Array{T, N}) where {T, N}
+@generated function _cubemap(input::Array{T,N}) where {T,N}
     quote
         result = similar(input, size(input) .* 2 .- 1)
         result .= typemin(T)
@@ -27,10 +27,10 @@
     end
 end
 
-_one_hot(i, ::Val{N}) where N = CartesianIndex{N}(ntuple(isequal(i), Val(N)))
+_one_hot(i, ::Val{N}) where {N} = CartesianIndex{N}(ntuple(isequal(i), Val(N)))
 
 # Convenience functions from converting cubemap index to vertices and back.
-@inline function _from_cubemap(root::CartesianIndex{K}, ::Val{N}) where {K, N}
+@inline function _from_cubemap(root::CartesianIndex{K}, ::Val{N}) where {K,N}
     2^count(iseven, Tuple(root)) == N || throw(ArgumentError("invalid N"))
     result = ntuple(_ -> root, Val(N))
     for (i, j) in enumerate(Tuple(root))
@@ -39,7 +39,7 @@ _one_hot(i, ::Val{N}) where N = CartesianIndex{N}(ntuple(isequal(i), Val(N)))
                 ifelse(isodd(k), 1, -1) * _one_hot(i, Val(K))
             end
         end
-        result = TupleTools.sort(result, by=Base.Fix2(getindex, i))
+        result = TupleTools.sort(result; by=Base.Fix2(getindex, i))
     end
     result = SVector(TupleTools.sort(result))
     return map(result) do c
@@ -47,7 +47,7 @@ _one_hot(i, ::Val{N}) where N = CartesianIndex{N}(ntuple(isequal(i), Val(N)))
     end
 end
 
-function _to_cubemap(vertices::NTuple{N}) where N
+function _to_cubemap(vertices::NTuple{N}) where {N}
     K = length(first(vertices))
     result = ntuple(Val(K)) do i
         sum(2 .* getindex.(vertices, i) .- 1) ÷ N
@@ -58,7 +58,7 @@ end
 function _is_valid(vertices::Tuple{})
     return true
 end
-function _is_valid(vertices::NTuple{N}) where N
+function _is_valid(vertices::NTuple{N}) where {N}
     floor(log2(N)) == log2(N) || return false
     allunique(vertices) || return false
     K = length(first(vertices))
@@ -85,19 +85,19 @@ julia> Cube{1}(CartesianIndex(1, 2), 1.0)
 
 ```
 """
-struct Cube{D, T, K} <: AbstractSimplex{D, T, CartesianIndex{K}}
-    root::NTuple{K, Int32}
+struct Cube{D,T,K} <: AbstractSimplex{D,T,CartesianIndex{K}}
+    root::NTuple{K,Int32}
     birth::T
 
-    function Cube{D, T, K}(root::CartesianIndex{K}, birth) where {D, T, K}
+    function Cube{D,T,K}(root::CartesianIndex{K}, birth) where {D,T,K}
         D < 0 && throw(DomainError("dimension must be a non-negative integer"))
-        return new{D, T, K}(Int32.(Tuple(root)), convert(T, birth))
+        return new{D,T,K}(Int32.(Tuple(root)), convert(T, birth))
     end
 end
-function Cube{D}(root::CartesianIndex{K}, birth::T) where {D, T, K}
-    return Cube{D, T, K}(root, birth)
+function Cube{D}(root::CartesianIndex{K}, birth::T) where {D,T,K}
[…]
