+++ test/simplex.jl
@@ -4,30 +4,27 @@ using Test
 
 using Ripserer: _binomial, _vertices, coboundary, boundary
 
-struct FakeFiltration<:Ripserer.AbstractFiltration{Int, Int} end
+struct FakeFiltration <: Ripserer.AbstractFiltration{Int,Int} end
 function Ripserer.unsafe_simplex(
-    ::Type{Simplex{D, Int, Int}},
-    ::FakeFiltration,
-    vertices,
-    sign,
-) where D
-    return Simplex{D, Int, Int}(sign * index(vertices), 1)
+    ::Type{Simplex{D,Int,Int}}, ::FakeFiltration, vertices, sign
+) where {D}
+    return Simplex{D,Int,Int}(sign * index(vertices), 1)
 end
 Ripserer.nv(::FakeFiltration) = 20
-Ripserer.simplex_type(::Type{FakeFiltration}, D) = Simplex{D, Int, Int}
+Ripserer.simplex_type(::Type{FakeFiltration}, D) = Simplex{D,Int,Int}
 
-struct FakeFiltrationWithThreshold<:Ripserer.AbstractFiltration{Int, Int} end
+struct FakeFiltrationWithThreshold <: Ripserer.AbstractFiltration{Int,Int} end
 function Ripserer.unsafe_simplex(
-    ::Type{Simplex{D, Int, Int}}, ::FakeFiltrationWithThreshold, vertices, sign
-) where D
+    ::Type{Simplex{D,Int,Int}}, ::FakeFiltrationWithThreshold, vertices, sign
+) where {D}
     if maximum(vertices) > 10
         return nothing
     else
-        return Simplex{D, Int, Int}(sign * index(vertices), 1)
+        return Simplex{D,Int,Int}(sign * index(vertices), 1)
     end
 end
 Ripserer.nv(::FakeFiltrationWithThreshold) = 20
-Ripserer.simplex_type(::Type{FakeFiltrationWithThreshold}, D) = Simplex{D, Int, Int}
+Ripserer.simplex_type(::Type{FakeFiltrationWithThreshold}, D) = Simplex{D,Int,Int}
 
 @testset "Internal functions" begin
     @testset "_binomial" begin
@@ -38,19 +35,22 @@ Ripserer.simplex_type(::Type{FakeFiltrationWithThreshold}, D) = Simplex{D, Int, 
         @test index((3, 2, 1)) === 1
         @test index(Int32.((4, 3, 2, 1))) === Int32(1)
         @test index(SVector(101, 100, 20, 15, 3)) ==
-            1 +
-            binomial(101 - 1, 5) +
-            binomial(100 - 1, 4) +
-            binomial(20 - 1, 3) +
-            binomial(15 - 1, 2) +
-            binomial(3 - 1, 1)
+              1 +
+              binomial(101 - 1, 5) +
+              binomial(100 - 1, 4) +
+              binomial(20 - 1, 3) +
+              binomial(15 - 1, 2) +
+              binomial(3 - 1, 1)
     end
 
     @testset "_vertices" begin
         for i in 1:3:10, N in 1:5
             @test index(_vertices(i, Val(N))) == i
             @test length(_vertices(i, Val(N))) == N
-            @test begin @inferred _vertices(i, Val(N)); true end
+            @test begin
+                @inferred _vertices(i, Val(N))
+                true
+            end
         end
     end
 end
@@ -63,17 +63,15 @@ end
     end
 
     @testset "AbstractSimplex interface" begin
-        for D in (0, 2, 4),
-            T in (Float64, Float32, Int),
-            I in (Int32, Int64, Int128)
+        for D in (0, 2, 4), T in (Float64, Float32, Int), I in (Int32, Int64, Int128)
[…]
