+++ test/filtrations/test-datasets.jl
@@ -13,64 +13,70 @@ function rand_dist_matrix(n)
     A = rand(n, n)
     A .+= A'
     A -= Diagonal(A)
-    A
+    return A
 end
 function rand_dist_matrix(n, sparse)
-    A = sprand(n, n, sparse/2)
+    A = sprand(n, n, sparse / 2)
     A .+= A'
     A -= Diagonal(A)
     dropzeros!(A)
-    A
+    return A
 end
 
 "Distances on an icosahedron graph with edge length 1."
-icosahedron = Float64[0 1 2 2 1 2 1 1 2 2 1 3;
-                      1 0 3 2 1 1 2 1 2 1 2 2;
-                      2 3 0 1 2 2 1 2 1 2 1 1;
-                      2 2 1 0 3 2 1 1 2 1 2 1;
-                      1 1 2 3 0 1 2 2 1 2 1 2;
-                      2 1 2 2 1 0 3 2 1 1 2 1;
-                      1 2 1 1 2 3 0 1 2 2 1 2;
-                      1 1 2 1 2 2 1 0 3 1 2 2;
-                      2 2 1 2 1 1 2 3 0 2 1 1;
-                      2 1 2 1 2 1 2 1 2 0 3 1;
-                      1 2 1 2 1 2 1 2 1 3 0 2;
-                      3 2 1 1 2 1 2 2 1 1 2 0]
+icosahedron = Float64[
+    0 1 2 2 1 2 1 1 2 2 1 3
+    1 0 3 2 1 1 2 1 2 1 2 2
+    2 3 0 1 2 2 1 2 1 2 1 1
+    2 2 1 0 3 2 1 1 2 1 2 1
+    1 1 2 3 0 1 2 2 1 2 1 2
+    2 1 2 2 1 0 3 2 1 1 2 1
+    1 2 1 1 2 3 0 1 2 2 1 2
+    1 1 2 1 2 2 1 0 3 1 2 2
+    2 2 1 2 1 1 2 3 0 2 1 1
+    2 1 2 1 2 1 2 1 2 0 3 1
+    1 2 1 2 1 2 1 2 1 3 0 2
+    3 2 1 1 2 1 2 2 1 1 2 0
+]
 
 "Distances on a cycle graph with edge length 1."
-cycle = [0 1 2 3 4 5 6 7 8 9 8 7 6 5 4 3 2 1;
-         1 0 1 2 3 4 5 6 7 8 9 8 7 6 5 4 3 2;
-         2 1 0 1 2 3 4 5 6 7 8 9 8 7 6 5 4 3;
-         3 2 1 0 1 2 3 4 5 6 7 8 9 8 7 6 5 4;
-         4 3 2 1 0 1 2 3 4 5 6 7 8 9 8 7 6 5;
-         5 4 3 2 1 0 1 2 3 4 5 6 7 8 9 8 7 6;
-         6 5 4 3 2 1 0 1 2 3 4 5 6 7 8 9 8 7;
-         7 6 5 4 3 2 1 0 1 2 3 4 5 6 7 8 9 8;
-         8 7 6 5 4 3 2 1 0 1 2 3 4 5 6 7 8 9;
-         9 8 7 6 5 4 3 2 1 0 1 2 3 4 5 6 7 8;
-         8 9 8 7 6 5 4 3 2 1 0 1 2 3 4 5 6 7;
-         7 8 9 8 7 6 5 4 3 2 1 0 1 2 3 4 5 6;
-         6 7 8 9 8 7 6 5 4 3 2 1 0 1 2 3 4 5;
-         5 6 7 8 9 8 7 6 5 4 3 2 1 0 1 2 3 4;
[…]
