+++ docs/make.jl
@@ -10,17 +10,18 @@ ENV["GKSwstype"] = "100"
 EXAMPLES_INPUT = joinpath(@__DIR__, "src/examples")
 EXAMPLES_OUTPUT = joinpath(@__DIR__, "src/generated")
 
-for example in readdir(EXAMPLES_INPUT, join=true)
+for example in readdir(EXAMPLES_INPUT; join=true)
     endswith(example, ".jl") || continue
-    Literate.markdown(example, EXAMPLES_OUTPUT, documenter=true)
+    Literate.markdown(example, EXAMPLES_OUTPUT; documenter=true)
 end
 
-makedocs(
+makedocs(;
     sitename="Ripserer.jl",
-    format = Documenter.HTML(
-        # Use clean URLs, unless built as a "local" build
-        prettyurls = !("local" in ARGS),
-        assets = ["assets/favicon.ico"],
+    format=Documenter.HTML(
+        # Use clean URLs, unless built as a "local" build;
+        ;
+        prettyurls=!("local" in ARGS),
+        assets=["assets/favicon.ico"],
     ),
     pages=[
         "Home" => "index.md",
@@ -32,16 +33,11 @@ makedocs(
             "generated/malaria.md",
         ],
         "Benchmarks" => "benchmarks.md",
-        "API" => [
-            "Public" => "api/ripserer.md",
-            "Interfaces" => "api/extensions.md",
-        ],
+        "API" => ["Public" => "api/ripserer.md", "Interfaces" => "api/extensions.md"],
         "Related Julia Packages" => "related-work.md",
         "Acknowledgements and References" => "references.md",
     ],
     doctest=false, # Doctests are run as part of testing -- no need to run them twice.
 )
 
-deploydocs(
-    repo = "github.com/mtsch/Ripserer.jl.git",
-)
+deploydocs(; repo="github.com/mtsch/Ripserer.jl.git")

[…]
