### 2020-10-17 bf10bae
Formatting (#117)
* use JuliaFormatter.jl

* formatter CI

* format tests

* format docs

* fix cubical docs

* format benchmarks
