+++ test/cycles.jl
@@ -23,7 +23,7 @@ using Ripserer: distance_matrix, OneSkeleton
         @test has_vertex(g, 6)
         @test !has_vertex(g, 7)
         @test !has_vertex(g, 0)
-        @test edges(g) == [Edge(j, i) for i in 1:5 for j in i+1:6]
+        @test edges(g) == [Edge(j, i) for i in 1:5 for j in (i + 1):6]
         @test vertices(g) == 1:6
         @test is_directed(g) == is_directed(typeof(g)) == false
         @test weights(g) == distance_matrix(flt)
@@ -52,8 +52,10 @@ using Ripserer: distance_matrix, OneSkeleton
         @test has_vertex(g, 6)
         @test !has_vertex(g, 7)
         @test !has_vertex(g, 0)
-        @test edges(g) == [Edge(j, i) for i in 1:5 for j in i+1:6
-                           if dists[i, j] ≠ 3 && (j, i) ∉ [(2, 1), (3, 1)]]
+        @test edges(g) == [
+            Edge(j, i) for i in 1:5
+            for j in (i + 1):6 if dists[i, j] ≠ 3 && (j, i) ∉ [(2, 1), (3, 1)]
+        ]
         @test vertices(g) == 1:6
         @test is_directed(g) == is_directed(typeof(g)) == false
         @test weights(g) == distance_matrix(flt)
@@ -65,10 +67,12 @@ using Ripserer: distance_matrix, OneSkeleton
     end
 
     @testset "simplex as threshold" begin
-        dists = [0 1 1 1
-                 1 0 1 1
-                 1 1 0 1
-                 1 1 1 0]
+        dists = [
+            0 1 1 1
+            1 0 1 1
+            1 1 0 1
+            1 1 1 0
+        ]
         flt = Rips(dists)
         g = OneSkeleton(flt, Simplex{1}((3, 2), 1))
 
@@ -79,7 +83,7 @@ using Ripserer: distance_matrix, OneSkeleton
         @test has_vertex(g, 4)
         @test !has_vertex(g, 7)
         @test !has_vertex(g, 0)
-        @test edges(g) == [Edge(j, i) for i in 1:3 for j in i+1:4 if (j, i) ≥ (3, 2)]
+        @test edges(g) == [Edge(j, i) for i in 1:3 for j in (i + 1):4 if (j, i) ≥ (3, 2)]
         @test vertices(g) == 1:4
         @test is_directed(g) == is_directed(typeof(g)) == false
         @test weights(g) == distance_matrix(flt)
@@ -130,9 +134,9 @@ end
             1 2 3 2 1 0
         ])
 
-        interval = ripserer(flt, reps=true)[2][1]
+        interval = ripserer(flt; reps=true)[2][1]
         cyc1 = reconstruct_cycle(flt, interval)
-        @test eltype(cyc1) == Simplex{1, Int, Int}
+        @test eltype(cyc1) == Simplex{1,Int,Int}
         @test all(birth.(cyc1) .== 1)
         @test allunique(cyc1)
         @test length(cyc1) == 6
@@ -142,12 +146,12 @@ end
     end
 
     @testset "circle points" begin
-        pts = [(sin(t), cos(t)) for t in range(0, 2π; length=22)[1:end-1]]
+        pts = [(sin(t), cos(t)) for t in range(0, 2π; length=22)[1:(end - 1)]]
         flt = Rips(pts)
-        interval = ripserer(flt, reps=true)[2][1]
+        interval = ripserer(flt; reps=true)[2][1]
 
         prev_length = nv(flt)
-        for r in range(birth(interval), death(interval); length=10)[1:end-1]
[…]
