+++ dev/examples/cubical.jl
@@ -7,27 +7,27 @@ using Ripserer
 using Images
 using Plots
 gr() # hide
-using Random
-Random.seed!(1337)
+using Random # hide
+Random.seed!(1337) # hide
 nothing # hide
 
 # We will use two data sets in this example. The first will be a curve:
 
 n = 1000
-x = range(0, 1, length=n)
+x = range(0, 1; length=n)
 curve = sin.(2π * 5x) .* x
 
-curve_plot = plot(curve, legend=false, title="Curve")
+curve_plot = plot(curve; legend=false, title="Curve")
 
 # The other will be [the Event Horizon Telescope picture of a black
 # hole](https://en.wikipedia.org/wiki/File:Black_hole_-_Messier_87_crop_max_res.jpg). We
 # will use a small, 240×240 pixel version of the image. Ripserer should have no problems
 # with processing larger images, but this will work well enough for this tutorial.
 
-blackhole_image = load(
-    joinpath(@__DIR__, "../assets/data/240px-Black_hole_-_Messier_87_crop_max_res.jpg")
-)
-blackhole_plot = plot(blackhole_image, title="Black Hole")
+blackhole_image = load(joinpath(
+    @__DIR__, "../assets/data/240px-Black_hole_-_Messier_87_crop_max_res.jpg"
+))
+blackhole_plot = plot(blackhole_image; title="Black Hole")
 
 # To use the image with Ripserer, we have to convert it to grayscale.
 
@@ -48,7 +48,7 @@ plot(curve_plot, plot(result))
 # To see which minimum each interval corresponds to, we compute representatives by setting
 # `reps=true`.
 
-result, _ = ripserer(Cubical(curve), reps=true)
+result, _ = ripserer(Cubical(curve); reps=true)
 
 # The infinite interval's representative will include the whole function. To plot a
 # representative, simply pass the interval along with the data to `plot`. An alternative
@@ -56,13 +56,16 @@ result, _ = ripserer(Cubical(curve), reps=true)
 # plot that in the same way. Let's use both ways in the following example.
 
 infinite_interval = only(filter(!isfinite, result))
-plt = plot(representative(infinite_interval), curve,
-           legend=false,
-           title="Representatives",
-           seriestype=:path)
+plt = plot(
+    representative(infinite_interval),
+    curve;
+    legend=false,
+    title="Representatives",
+    seriestype=:path,
+)
 
 for interval in filter(isfinite, result)
-    plot!(plt, interval, curve, seriestype=:path)
+    plot!(plt, interval, curve; seriestype=:path)
 end
 
 # To get the locations of the minima, extract the critical simplices from intervals. Since
@@ -74,9 +77,8 @@ end
 min_indices = [only(birth_simplex(int)) for int in result]
 min_x = eachindex(curve)[min_indices]
 
-scatter!(plt, min_x, curve[min_x], color=1:6, markershape=:star)
-plot(plt, plot(result, markercolor=1:6, markeralpha=1))
-
+scatter!(plt, min_x, curve[min_x]; color=1:6, markershape=:star)
+plot(plt, plot(result; markercolor=1:6, markeralpha=1))
 
 # Note that each interval's birth is equal to the value of the corresponding local minimum
 # and its death is equal to the higher of the two adjacent maxima. An intuitive way of
@@ -94,15 +96,15 @@ plot(noisy_curve)
 # the diagram in birth, persistence coordinates, which will make the level of noise easier
 # to see.
 
-noisy_result, _ = ripserer(Cubical(noisy_curve), reps=true)
-plot(noisy_result, persistence=true)
+noisy_result, _ = ripserer(Cubical(noisy_curve); reps=true)
+plot(noisy_result; persistence=true)
 
[…]
