+++ dev/examples/basics.jl
@@ -4,7 +4,10 @@
 
 using Ripserer
 using Plots
-using Random; Random.seed!(1337); gr(); nothing # hide
+using Random # hide
+Random.seed!(1337) # hide
+gr() # hide
+nothing # hide
 
 # ## Basic Usage
 
@@ -12,28 +15,28 @@ using Random; Random.seed!(1337); gr(); nothing # hide
 # We start by defining our sampling function.
 
 function noisy_circle(n; r=1, noise=0.1)
-    points = NTuple{2, Float64}[]
+    points = NTuple{2,Float64}[]
     for _ in 1:n
         θ = 2π * rand()
-        push!(points, (r*sin(θ) + noise*rand(), r*cos(θ) + noise*rand()))
+        push!(points, (r * sin(θ) + noise * rand(), r * cos(θ) + noise * rand()))
     end
-    points
+    return points
 end
 
 # Next, we sample 100 points from the circle.
 
 circ_100 = noisy_circle(100)
-scatter(circ_100, aspect_ratio=1, legend=false, title="Noisy Circle")
+scatter(circ_100; aspect_ratio=1, legend=false, title="Noisy Circle")
 
 # To compute the persistent homology, simply run the following. The `dim_max` argument sets
 # the maximum dimension persistent homology is computed in.
 
-result_circ = ripserer(circ_100, dim_max=3)
+result_circ = ripserer(circ_100; dim_max=3)
 
 # !!! warning "Warning"
 #     Computing Vietoris-Rips persistent homology in high dimensions for large numbers of
 #     points is computationally expensive and requires a large amount of memory. Be careful
-#     or you **will** run out of memory. On an ordinary computer, you can expect to compute
+#     or you **will** run out of memory. On an ordinary laptop, you can expect to compute
 #     one-dimensional persistent homology for datasets of a few thousand points and higher
 #     (2-3) dimensional persistent homology for datasets of a few hundred points. This, of
 #     course, depends on the data set itself.
@@ -53,11 +56,11 @@ barcode(result_circ)
 #     `result` is just an array of persistence diagrams, so the zero-dimensional diagram is
 #     found at index 1.
 
-barcode(result_circ[2:end], linewidth=2)
+barcode(result_circ[2:end]; linewidth=2)
 
 # We can plot a single diagram in the same manner.
 
-barcode(result_circ[3], linewidth=3)
+barcode(result_circ[3]; linewidth=3)
 
 # ## Distance Matrix Inputs
 
@@ -70,23 +73,25 @@ barcode(result_circ[3], linewidth=3)
 # <img src="https://upload.wikimedia.org/wikipedia/commons/8/83/Icosahedron_graph.svg" height="250" width="250">
 # ```
 
-icosahedron = [0 1 2 2 1 2 1 1 2 2 1 3;
-               1 0 3 2 1 1 2 1 2 1 2 2;
-               2 3 0 1 2 2 1 2 1 2 1 1;
-               2 2 1 0 3 2 1 1 2 1 2 1;
-               1 1 2 3 0 1 2 2 1 2 1 2;
-               2 1 2 2 1 0 3 2 1 1 2 1;
-               1 2 1 1 2 3 0 1 2 2 1 2;
-               1 1 2 1 2 2 1 0 3 1 2 2;
-               2 2 1 2 1 1 2 3 0 2 1 1;
-               2 1 2 1 2 1 2 1 2 0 3 1;
-               1 2 1 2 1 2 1 2 1 3 0 2;
-               3 2 1 1 2 1 2 2 1 1 2 0]
-nothing; # hide
+icosahedron = [
+    0 1 2 2 1 2 1 1 2 2 1 3
[…]
