### 2020-10-19 #2614 @VivianePons
Thank you all for the great work! The paper and software have received many praises, it is pleasure to accept it into JOSS, thank you @mtsch for the work and thank you @juliohm and @sauln for the reviews :)
